% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_geom.R
\name{add_geom}
\alias{add_geom}
\title{add_geom}
\usage{
add_geom(data, unit, variable, crs = 32721)
}
\arguments{
\item{data}{data frame con las variables necesarias de ech.}

\item{unit}{unidad espacial de la base "Departamentos", "Secciones", "Secc MVD 2004", "Segmentos", "Segm MVD 2004", "Segm URB INT 2004", "Zonas", "Zonas MVD 2004", "Zonas URB INT 2004", "Localidades pg", "Municipios" o "Barrios".}

\item{variable}{nombre de la variable con el codigo correspondiente sin duplicados.}

\item{crs}{sistema de coordenadas de referencia, los màs usados para la region 32721 and 4326 (default 32721)}
}
\value{
data.frame
}
\description{
This function allows you to add geom variable to a with a code varible of "zona", "barrio", "localidad", "segmentos", "secciones" or "departamentos".
}
\details{
Disclaimer: El script no es un producto oficial de INE.
}
