% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{age_groups}
\alias{age_groups}
\title{age_groups}
\usage{
age_groups(
  data = ech::toy_ech_2018,
  cut = c(0, 4, 11, 17, 24),
  e27 = "e27",
  labels = c("0-4", "5-11", "12-17", "18-24", "+24")
)
}
\arguments{
\item{data}{data.frame}

\item{cut}{breaks points to cut a numeric variable}

\item{e27}{Variable name of age}

\item{labels}{Labels for the new variable}
}
\value{
data.frame
}
\description{
age_groups
}
\examples{
df <- age_groups(data = ech::toy_ech_2018, cut = c(0, 4, 11, 17, 24))

}
