% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_microdata.R
\name{archive_extract}
\alias{archive_extract}
\title{archive_extract}
\usage{
archive_extract(archive.path = NULL, dest.path = NULL)
}
\arguments{
\item{archive.path}{Ruta de origen del archivo comprimido}

\item{dest.path}{Ruta destino del archivo descomprimido}
}
\description{
archive_extract
}
\details{
Disclaimer: This script is not an official INE product.
Aviso: El script no es un producto oficial de INE.
}
\examples{
\donttest{
archive_extract()
}
}
