% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{archive_read}
\alias{archive_read}
\title{archive_read ech from zip/rar file}
\usage{
archive_read(archive, file = 1L, mode = "r", format = NULL,
  filter = NULL)
}
\arguments{
\item{mode}{how to open the connection as used in \code{base::connections()}}
}
\description{
archive_read ech from zip/rar file
}
\keyword{internal}
