% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cba_cbna_int}
\alias{cba_cbna_int}
\title{Data CBA-CBNA (Interior Urbano)}
\format{
A data frame with 234 rows and 4 variables:
\describe{
  \item{fecha}{date from 2001 to 2020}
  \item{cba_li}{CBA}
  \item{cbna}{CBNA}
  \item{cbt_lp}{CBT}
 }
}
\source{
\url{http://www.ine.gub.uy/}
}
\usage{
cba_cbna_int
}
\description{
A dataset containing the CBA and CBNA for the Interior Urbano region
}
\keyword{datasets}
