% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cba_cbna_mdeo}
\alias{cba_cbna_mdeo}
\title{Data CBA-CBNA (Montevideo)}
\format{
A data frame with 234 rows and 4 variables:
\describe{
  \item{fecha}{date from 2001 to 2020}
  \item{cba_li}{CBA}
  \item{cbna}{CBNA}
  \item{cbt_lp}{CBT}
 }
}
\source{
\url{http://www.ine.gub.uy/}
}
\usage{
cba_cbna_mdeo
}
\description{
A dataset containing the CBA and CBNA for the Montevideo region
}
\details{
Disclaimer: This script is not an official INE product.
Aviso: El script no es un producto oficial de INE.
}
\keyword{datasets}
