% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ciiu4}
\alias{ciiu4}
\title{Data CIIU4 and 3}
\format{
A data frame with 738 rows and 3 variables:
\describe{
  \item{ciiu_4}{Code of Clasificaci贸n Industrial Internacional Uniforme Rev. 4}
  \item{description}{Description of Clasificaci贸n Industrial Internacional Uniforme Rev. 4}
  \item{ciiu_3}{Code of Clasificaci贸n Industrial Internacional Uniforme Rev. 3}
  ...
 }
}
\source{
\url{http://www.ine.gub.uy/}
}
\usage{
ciiu4
}
\description{
A dataset containing Clasificaci贸n Industrial Internacional Uniforme Rev. 4 and 3
}
\keyword{datasets}
