% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{deflate}
\alias{deflate}
\title{deflate}
\usage{
deflate(base.month = base.month, base.year = base.year, ipc = "country")
}
\arguments{
\item{base.month}{mes base}

\item{base.year}{anio base}

\item{ipc}{IPC a nivel nacional o Mdeo-Interior}
}
\description{
deflate
}
\examples{
\donttest{
deflate(base.month = 6, base.year = 2016)
}
}
