% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{deflate_gini}
\alias{deflate_gini}
\title{deflate_gini}
\usage{
deflate_gini(base_month = "01", base_year = "2005", ipc = "G", df_year = NULL)
}
\arguments{
\item{base_month}{mes base}

\item{base_year}{anio base}

\item{ipc}{IPC a nivel nacional ('G'), IPC para Montevideo ('M') e IPC para Interior ('I')}

\item{df_year}{anio del data frame}
}
\value{
data.frame
}
\description{
deflate_gini
}
\examples{
\donttest{
deflate_gini(base_month = "06", base_year = "2016", df_year = 2016)
}
}
