% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_microdata.R
\name{download_ech}
\alias{download_ech}
\title{Download ECH from INE website}
\usage{
download_ech(years = NULL, folder = getwd(), format = "sav")
}
\arguments{
\item{years}{si no se indica un anio, descarga todos los anios disponibles}

\item{folder}{si no se indica una carpeta, descarga en la carpeta de
trabajo o la raiz del proyecto}

\item{format}{si no se indica se descarga en formato .sav (SPSS), la otra
opción sería "dat"}
}
\value{
los archivos comprimidos de la encuesta ECH descargados desde el
sitio web del INE
}
\description{
Download ECH from INE website
}
\examples{
# descargar todas las encuestas disponibles
# download_ech()

# descargar ECH 2017 en carpeta especifica
# download_ech(2017, "data-raw")
}
