% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/employment.R
\name{employment}
\alias{employment}
\title{An employment function}
\usage{
employment(data = ech::toy_ech_2018, pobpcoac = "pobpcoac")
}
\arguments{
\item{data}{data frame with microdata}

\item{pobpcoac}{Definition of population by activity status}
}
\description{
This function allows you to calculate the variables: PEA, PET, PO, PD
}
\details{
Disclaimer: El script no es un producto oficial de INE.
}
\examples{
\donttest{
toy_ech_2018 <- employment(data = ech::toy_ech_2018, pobpcoac = "pobpcoac")
}
}
\keyword{employment}
