% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/employment_rate.R
\name{employment_rate}
\alias{employment_rate}
\title{An employment Function}
\usage{
employment_rate(data = df, geo.unit = "uy", by = NULL,
  pobcoac = "pobcoac", weights = "pesoano")
}
\arguments{
\item{data}{data frame with microdata}
}
\description{
This function allows you to..
}
\examples{
employment_rate()
}
\keyword{employment}
