% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/employment.R
\name{employment_restrictions}
\alias{employment_restrictions}
\title{employment_restrictions}
\usage{
employment_restrictions(
  data = ech::toy_ech_2018,
  f82 = "f82",
  underemployment = "underemployment"
)
}
\arguments{
\item{data}{data.frame}

\item{f82}{Variable name of}

\item{underemployment}{Variable name of}
}
\value{
data.frame
}
\description{
employment_restrictions
}
