% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrolled_school.R
\name{enrolled_school}
\alias{enrolled_school}
\title{A function to estimate people enrolled in school}
\usage{
enrolled_school(data = ech::toy_ech_2018)
}
\arguments{
\item{data}{Base with necessary variables Defaults to ech.}
}
\description{
This function allows you to estimate people enrolled in school
}
\examples{
edu.asist(df)
}
\keyword{education}
