% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_ciiu}
\alias{get_ciiu}
\title{Title}
\usage{
get_ciiu(folder = tempdir(), version = 4)
}
\arguments{
\item{folder}{temp folder}

\item{version}{by default the last ciiu version}
}
\value{

}
\description{
Title
}
\examples{
\donttest{
get_ciiu(folder = tempdir())
}
}
