% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_estimation.R
\name{get_estimation}
\alias{get_estimation}
\title{A function to estimate variables at universe level}
\usage{
get_estimation(data = df, variable = NULL, by.x = NULL,
  by.y = NULL, level = NULL, ids = 1, strata = strata,
  weights = weights)
}
\arguments{
\item{data}{data frame with ECH microdata}

\item{variable}{data frame column to estimate}

\item{by.x}{data frame column}

\item{by.y}{data frame column}

\item{level}{is household or individual}

\item{ids}{survey phase}

\item{strata}{survey strata}

\item{weights}{ponderation variable}

\item{domain}{subpopulation reference}
}
\value{
table
}
\description{
This function allows you to estimate variable.
}
\examples{
get_estimation(data = d)
}
\keyword{household_type}
