% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_estimation.R
\name{get_estimation_gini}
\alias{get_estimation_gini}
\title{get_estimation_gini}
\usage{
get_estimation_gini(
  data = ech::toy_ech_2018,
  variable = y_wrv_pc_d_r,
  by.x = NULL,
  by.y = NULL,
  domain = NULL,
  level = NULL,
  name = "estimacion"
)
}
\arguments{
\item{data}{ech data frame}

\item{variable}{income without rental value per capita deflated}

\item{by.x}{variable}

\item{by.y}{variable}

\item{domain}{subpoblacion}

\item{level}{household or individual}

\item{name}{nombre}
}
\value{
table
}
\description{
get_estimation_gini
}
\examples{
\donttest{
toy_ech_18 <- gini_income(data = ech::toy_ech_2018)
get_estimation_gini(data = toy_ech_18, level = "h")
}
}
