% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_estimation.R
\name{get_estimation_total}
\alias{get_estimation_total}
\title{get_estimation_total}
\usage{
get_estimation_total(
  data = ech::toy_ech_2018,
  variable = NULL,
  by.x = NULL,
  by.y = NULL,
  domain = NULL,
  level = NULL,
  ids = NULL,
  numero = "numero",
  estred13 = "estred13",
  pesoano = "pesoano",
  name = "estimacion"
)
}
\arguments{
\item{data}{data frame with ECH microdata}

\item{variable}{data frame column to estimate}

\item{by.x}{data frame column}

\item{by.y}{data frame column}

\item{domain}{subpopulation reference setted as logical}

\item{level}{is household ("h") or individual ("i").}

\item{ids}{ids}

\item{numero}{household id}

\item{estred13}{strata}

\item{pesoano}{weights}

\item{name}{name for the estimation new column}
}
\value{
table
}
\description{
This function allows you to estimate total variable at universe level.
}
\details{
Disclaimer: This script is not an official INE product.
Aviso: El script no es un producto oficial de INE.
}
\examples{
\donttest{
get_estimation_total(variable = "pobre06", by.x = "dpto", level = "h")
}
}
\keyword{inference}
