% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_ipab}
\alias{get_ipab}
\title{get_ipab}
\usage{
get_ipab(folder = tempdir(), sheet = NULL)
}
\arguments{
\item{folder}{temporal folder}

\item{sheet}{number of sheet}
}
\value{
data.frame
}
\description{
IPAB (Indice de precios de alimentos y bebidas)
}
\examples{
get_ipab(folder = tempdir(), sheet = 1)

}
