% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_ipab_region}
\alias{get_ipab_region}
\title{This function allows you to get the IPAB (Indice de precios de alimentos y bebidas) data}
\usage{
get_ipab_region(folder = tempdir(), sheet = NULL, region = "M")
}
\arguments{
\item{folder}{temporal folder}

\item{sheet}{sheet number}

\item{region}{Montevideo ("M"), Interior Urbano ("I")}
}
\value{
data.frame
}
\description{
This function allows you to get the IPAB (Indice de precios de alimentos y bebidas) data
}
\examples{
get_ipab_region(folder = tempdir(), sheet = 1, region = "M")

}
