% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_ipc}
\alias{get_ipc}
\title{get_ipc}
\usage{
get_ipc(folder = tempdir())
}
\arguments{
\item{folder}{ruta temporal para descargar el archivo}
}
\description{
get_ipc
}
\details{
Disclaimer: This script is not an official INE product.
Aviso: El script no es un producto oficial de INE.
}
\examples{
\donttest{
get_ipc(folder = tempdir())
}
}
