% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_ipc_region}
\alias{get_ipc_region}
\title{Title}
\usage{
get_ipc_region(folder = tempdir(), region = "M", sheet = NULL)
}
\arguments{
\item{folder}{ruta temporal para descargar el archivo}

\item{region}{Montevideo ("M") o Interior ("I")}

\item{sheet}{numero de hoja de la planilla}
}
\value{
data.frame
}
\description{
Title
}
\examples{
\donttest{
get_ipc_region(folder = tempdir(), region = "M")
}

}
