% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_microdata.R
\name{get_microdata}
\alias{get_microdata}
\title{Descarga y lee los archivos de la ECH}
\usage{
get_microdata(years = NULL, folder = tempdir(), format = "sav")
}
\arguments{
\item{years}{si no se indica un anio, descarga todos los anios disponibles}

\item{folder}{si no se indica una carpeta, descarga en la carpeta temp}

\item{format}{si no se indica se descarga en formato .sav (SPSS), la otra
opción sería "dat"}
}
\value{
la encuesta ECH en formato tibble
}
\description{
Descarga y lee los archivos de la ECH
}
\examples{
# Obtener la ECH 2017
get_microdata(2017)
}
