% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/income.R
\name{gini_income}
\alias{gini_income}
\title{gini income}
\usage{
gini_income(
  data = ech::toy_ech_2018,
  base_month = "01",
  base_year = "2005",
  mes = "mes"
)
}
\arguments{
\item{data}{data frame with ECH microdata}

\item{base_month}{month base}

\item{base_year}{year base}

\item{mes}{mes}
}
\description{
gini income
}
\details{
Disclaimer: El script no es un producto oficial de INE.
}
