% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/household_type.R
\name{household_type}
\alias{household_type}
\title{A household function}
\usage{
household_type(
  data = ech::toy_ech_2018,
  e26 = "e26",
  e27 = "e27",
  e30 = "e30",
  colname = "tipo_hogar"
)
}
\arguments{
\item{data}{data frame with ECH microdata}

\item{e26}{data frame column ('sex')}

\item{e27}{data frame column ('age')}

\item{e30}{data frame column ('householder')}

\item{colname}{custom name for the new variable}
}
\description{
This function allows you to calculate the household type for each household in the survey. A household is composed of one or more people who occupy a housing unit.
}
\examples{
\donttest{
household_type(data = ech::toy_ech_2018)
}
}
\keyword{household_type}
