% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwelling.R
\name{housing_conditions}
\alias{housing_conditions}
\title{A function to calculate the housing conditions}
\usage{
housing_conditions(data = ech::toy_ech_2018, c2 = "c2", c3 = "c3", c4 = "c4")
}
\arguments{
\item{data}{data.frame}

\item{c2}{predominant material on external walls}

\item{c3}{predominant roofing material}

\item{c4}{predominant flooring material}
}
\value{
data.frame
}
\description{
A function to calculate the housing conditions
}
\examples{
toy_ech_2018 <- housing_conditions(data = ech::toy_ech_2018)

}
