% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwelling.R
\name{housing_deprivation}
\alias{housing_deprivation}
\title{housing_deprivation}
\usage{
housing_deprivation(
  data = ech::toy_ech_2018,
  ht19 = "ht19",
  d9 = "d9",
  d10 = "d10",
  d11 = "d11",
  d12 = "d12",
  d13 = "d13",
  d16 = "d16",
  d18 = "d18",
  d19 = "d19",
  c2 = "c2",
  c3 = "c3",
  c4 = "c4",
  quintil = "quintil",
  region_4 = "region_4"
)
}
\arguments{
\item{data}{data frame}

\item{ht19}{number of individuals in the household}

\item{d9}{number of rooms}

\item{d10}{number of rooms to sleep}

\item{d11}{source of water}

\item{d12}{llegada del agua a la vivienda}

\item{d13}{servicio sanitario}

\item{d16}{evacuacion del servicio sanitario}

\item{d18}{energy source for lighting}

\item{d19}{cooking space}

\item{c2}{predominant material on external walls}

\item{c3}{predominant roofing material}

\item{c4}{predominant flooring material}

\item{quintil}{income quintil}

\item{region_4}{region}
}
\value{
data.frame
}
\description{
This function allows you to calculate the housing status
}
\details{
Disclaimer: El script no es un producto oficial de INE.
}
\examples{
\donttest{
toy_ech_2018 <- income_constant_prices(data = ech::toy_ech_2018)
toy_ech_2018 <- income_quantiles(data = toy_ech_2018)
toy_ech_2018 <- housing_deprivation(data = toy_ech_2018)
}
}
