% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwelling.R
\name{housing_situation}
\alias{housing_situation}
\title{A function to calculate the housing situation}
\usage{
housing_situation(
  data = ech::toy_ech_2018,
  c5_1 = "c5_1",
  c5_2 = "c5_2",
  c5_3 = "c5_3",
  c5_4 = "c5_4",
  c5_5 = "c5_5",
  c5_6 = "c5_6",
  c5_7 = "c5_7",
  c5_8 = "c5_8",
  c5_9 = "c5_9",
  c5_10 = "c5_10",
  c5_11 = "c5_11",
  c5_12 = "c5_12"
)
}
\arguments{
\item{data}{data.frame}

\item{c5_1}{roof condensation}

\item{c5_2}{roof drips}

\item{c5_3}{walls cracks}

\item{c5_4}{broken doors or windows}

\item{c5_5}{floors cracks}

\item{c5_6}{plaster drop on walls}

\item{c5_7}{detached ceilings}

\item{c5_8}{poor sunlight}

\item{c5_9}{poor ventilation}

\item{c5_10}{floods when it rains}

\item{c5_11}{in danger of collapse}

\item{c5_12}{dampness in the foundations}
}
\value{
data.frame
}
\description{
A function to calculate the housing situation
}
\examples{
toy_ech_2018 <- housing_situation(data = ech::toy_ech_2018)

}
