% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwelling.R
\name{housing_tenure}
\alias{housing_tenure}
\title{A function to calculate the housing tenure}
\usage{
housing_tenure(data = ech::toy_ech_2018, d8_1 = "d8_1")
}
\arguments{
\item{data}{data.frame}

\item{d8_1}{housing_tenure (owner, renter, rent-free occupancy, etc.)}
}
\value{
data.frame
}
\description{
A function to calculate the housing tenure
}
\examples{
toy_ech_2018 <- housing_tenure(data = ech::toy_ech_2018)

}
