% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/income.R
\name{income_quantiles}
\alias{income_quantiles}
\title{income_quantiles}
\usage{
income_quantiles(
  data = ech::toy_ech_2018,
  quantile = 5,
  weights = "pesoano",
  income = "y_pc_d"
)
}
\arguments{
\item{data}{data.frame}

\item{quantile}{cuantiles: quintiles (5) o deciles (10)}

\item{weights}{ponderation variable}

\item{income}{Name of the income variables. Default: "y_pc_d"}
}
\value{
data.frame
}
\description{
Household Income Quantiles
}
\details{
Disclaimer: El script no es un producto oficial de INE.
}
\examples{
\donttest{
toy_ech_2018 <- income_constant_prices(data = ech::toy_ech_2018)
toy_ech_2018 <- income_quantiles(data = toy_ech_2018)
}
}
