% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/income_quantiles.R
\name{income_quintiles}
\alias{income_quintiles}
\title{income_quintiles}
\usage{
income_quintiles(
  data = df,
  ht11_per_capita_deflate = ht11_per_capita_deflate,
  quantile = 5,
  weights = pesoano
)
}
\arguments{
\item{data}{data.frame}

\item{ht11_per_capita_deflate}{ingeso per capita deflactado}

\item{quantile}{cuantiles}

\item{weights}{ponderador de los casos}
}
\value{

}
\description{
Household Income Quantiles
}
