% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poverty.R
\name{integrated_poverty_measure}
\alias{integrated_poverty_measure}
\title{This function allows you to calculate an integrated poverty measure}
\usage{
integrated_poverty_measure(
  data = ech::toy_ech_2018,
  pobre06 = "pobre06",
  UBN_q = "UBN_q"
)
}
\arguments{
\item{data}{data.frame}

\item{pobre06}{Variable name of poverty}

\item{UBN_q}{Variable name of UBN}
}
\value{
data.frame
}
\description{
This function allows you to calculate an integrated poverty measure
}
\examples{
\donttest{
toy_ech_18 <- enrolled_school(data = ech::toy_ech_2018)
toy_ech_18 <- years_of_schooling(toy_ech_18)
toy_ech_18 <- unsatisfied_basic_needs(toy_ech_18)
toy_ech_18 <- integrated_poverty_measure(data = toy_ech_18)
}
}
\seealso{
Other poverty: 
\code{\link{poverty}()},
\code{\link{unsatisfied_basic_needs}()}
}
\concept{poverty}
