% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ipab_base2010}
\alias{ipab_base2010}
\title{Data IPAB}
\format{
A data frame with 286 rows and 3 variables:
\describe{
  \item{yy}{date from 1997 to 2020}
  \item{mm}{month}
  \item{indice}{IPAB}
 }
}
\source{
\url{http://www.ine.gub.uy/}
}
\usage{
ipab_base2010
}
\description{
A dataset containing the IPAB
}
\keyword{datasets}
