% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ipc_base2010}
\alias{ipc_base2010}
\title{Data IPC}
\format{
A data frame with 990 rows and 5 variables:
\describe{
  \item{fecha}{date from 1937 to 2019}
  \item{indice}{IPC}
  \item{mensual}{value of IPC}
  \item{acum_ano}{acumulated IPC}
  \item{acum_12_meses}{acumulated IPC last 12 month}
 }
}
\source{
\url{http://www.ine.gub.uy/}
}
\usage{
ipc_base2010
}
\description{
A dataset containing the IPC base 2010
}
\keyword{datasets}
