% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ipc_base2010_mdeo}
\alias{ipc_base2010_mdeo}
\title{Data IPC (Montevideo)}
\format{
A data frame with 120 rows and 2 variables:
\describe{
  \item{fecha}{date from 2011 to 2019}
  \item{indice}{IPC}
 }
}
\source{
\url{http://www.ine.gub.uy/}
}
\usage{
ipc_base2010_mdeo
}
\description{
A dataset containing the IPC base 2010 only for the Montevideo region
}
\keyword{datasets}
