% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_microdata.R
\name{isWindows}
\alias{isWindows}
\title{This is used so that unit tests can override this using \code{testthat::with_mock}.}
\usage{
isWindows()
}
\description{
This is used so that unit tests can override this using \code{testthat::with_mock}.
}
\keyword{internal}
