% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/income.R
\name{labor_income_per_capita}
\alias{labor_income_per_capita}
\title{labor_income_per_capita}
\usage{
labor_income_per_capita(
  data = ech::toy_ech_2018,
  numero = "numero",
  pobpcoac = "pobpcoac",
  g126_1 = "g126_1",
  g126_2 = "g126_2",
  g126_3 = "g126_3",
  g126_4 = "g126_4",
  g126_5 = "g126_5",
  g126_6 = "g126_6",
  g126_7 = "g126_7",
  g126_8 = "g126_8",
  g127_3 = "g127_3",
  g128_1 = "g128_1",
  g129_2 = "g129_2",
  g130_1 = "g130_1",
  g131_1 = "g131_1",
  g133_1 = "g133_1",
  g133_2 = "g133_2",
  g134_1 = "g134_1",
  g134_2 = "g134_2",
  g134_3 = "g134_3",
  g134_4 = "g134_4",
  g134_5 = "g134_5",
  g134_6 = "g134_6",
  g134_7 = "g134_7",
  g134_8 = "g134_8",
  g135_3 = "g135_3",
  g136_1 = "g136_1",
  g137_2 = "g137_2",
  g138_1 = "g138_1",
  g139_1 = "g139_1",
  g141_1 = "g141_1",
  g141_2 = "g141_2",
  g142 = "g142",
  g144_1 = "g144_1",
  g144_2_1 = "g144_2_1",
  g144_2_3 = "g144_2_3",
  g144_2_4 = "g144_2_4",
  g144_2_5 = "g144_2_5"
)
}
\arguments{
\item{data}{data frame}

\item{numero}{Variable name of household id}

\item{pobpcoac}{Variable name of definition of population by activity status}

\item{g126_1}{Variable name of sueldo o jornales liquidos}

\item{g126_2}{Variable name of comisiones, incentivos, horas extras, habilitaciones}

\item{g126_3}{Variable name of viaticos no sujetos a rendicion}

\item{g126_4}{Variable name of propinas}

\item{g126_5}{Variable name of aguinaldo}

\item{g126_6}{Variable name of salario vacacional}

\item{g126_7}{Variable name of pagos atrasados}

\item{g126_8}{Variable name of boletos de transporte}

\item{g127_3}{Variable name of recibio alimentos o bebidas}

\item{g128_1}{Variable name of recibio tickets alimentacion}

\item{g129_2}{Variable name of recibio vivienda o alojamiento}

\item{g130_1}{Variable name of recibio otro tipo de retribucion}

\item{g131_1}{Variable name of recibio otro tipo de complemento pagado por el/la empleador/a}

\item{g133_1}{Variable name of derecho a cultivo para consumo propio}

\item{g133_2}{Variable name of derecho a cultivo para consumo propio (monto percibido por la venta)}

\item{g134_1}{Variable name of sueldo o jornales liquidos}

\item{g134_2}{Variable name of comisiones, incentivos, horas extras, habilitaciones}

\item{g134_3}{Variable name of viaticos no sujetos a rendicion}

\item{g134_4}{Variable name of propinas}

\item{g134_5}{Variable name of aguinaldo}

\item{g134_6}{Variable name of salario vacacional}

\item{g134_7}{Variable name of pagos atrasados}

\item{g134_8}{Variable name of boletos de transporte}

\item{g135_3}{Variable name of recibio alimentos o bebidas}

\item{g136_1}{Variable name of recibio tickets alimentacion}

\item{g137_2}{Variable name of recibio vivienda o alojamiento}

\item{g138_1}{Variable name of recibio otro tipo de retribucion}

\item{g139_1}{Variable name of recibio otro tipo de complemento pagado por el/la empleador/a}

\item{g141_1}{Variable name of derecho a cultivo para consumo propio}

\item{g141_2}{Variable name of derecho a cultivo para consumo propio (monto percibido por la venta)}

\item{g142}{Variable name of retiros para gastos del hogar de negocios que tiene o tenia}

\item{g144_1}{Variable name of retiro de productos para consumo propio (trabajador/a no agropecuario/a)}

\item{g144_2_1}{Variable name of retiro de productos para consumo propio (trabajador/a no agropecuario/a)}

\item{g144_2_3}{Variable name of retiro de productos para consumo propio (trabajador/a no agropecuario/a)}

\item{g144_2_4}{Variable name of retiro de productos para consumo propio (trabajador/a no agropecuario/a)}

\item{g144_2_5}{Variable name of retiro de productos para consumo propio (trabajador/a no agropecuario/a)}
}
\value{
data.frame
}
\description{
labor_income_per_capita
}
\details{
Disclaimer: This script is not an official INE product.
Aviso: El script no es un producto oficial de INE.
}
\examples{
\donttest{
toy_ech_2018 <- labor_income_per_capita(data = ech::toy_ech_2018)
}

}
