% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/income.R
\name{labor_income_per_hour}
\alias{labor_income_per_hour}
\title{Title}
\usage{
labor_income_per_hour(
  data = ech::toy_ech_2018,
  numero = "numero",
  f85 = "f85",
  pobpcoac = "pobpcoac",
  pt4 = "pt4",
  base_month = "base.month",
  base_year = "base.year"
)
}
\arguments{
\item{data}{data frame}

\item{numero}{identificador de hogar}

\item{f85}{horas trabajadas por semana}

\item{pobpcoac}{condicion de actividad economica}

\item{pt4}{total de ingresos por trabajo}

\item{base_month}{mes base}

\item{base_year}{anio base}
}
\value{

}
\description{
Title
}
\details{
Disclaimer: El script no es un producto oficial de INE.
}
