% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/education.R
\name{level_completion}
\alias{level_completion}
\title{Title}
\usage{
level_completion(
  data = ech::toy_ech_2018,
  e197 = "e197",
  e197_1 = "e197_1",
  e201 = "e201",
  e51_4 = "e51_4",
  e51_5 = "e51_5",
  e51_6 = "e51_6"
)
}
\arguments{
\item{data}{ech}

\item{e197}{attends primary school}

\item{e197_1}{completed primary}

\item{e201}{attends secondary}

\item{e51_4}{years passed in lower secondary}

\item{e51_5}{years passed in upper secondary}

\item{e51_6}{years passed in bachillerato tecnologico}
}
\value{

}
\description{
Title
}
\details{
Disclaimer: El script no es un producto oficial de INE.
}
\examples{
\donttest{
toy_ech_2018 <- level_completion(data = ech::toy_ech_2018)
}
}
