% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/education.R
\name{level_education}
\alias{level_education}
\title{A function to calculate the highest level of education achieved}
\usage{
level_education(
  data = ech::toy_ech_2018,
  e51_2 = "e51_2",
  e51_3 = "e51_3",
  e51_4 = "e51_4",
  e51_5 = "e51_5",
  e51_6 = "e51_6",
  e51_7 = "e51_7",
  e51_7_1 = "e51_7_1",
  e51_8 = "e51_8",
  e51_9 = "e51_9",
  e51_10 = "e51_10",
  e51_11 = "e51_11",
  e193 = "e193",
  e49 = "e49"
)
}
\arguments{
\item{data}{data frame}

\item{e51_2}{years passed in primary}

\item{e51_3}{years passed in special primary}

\item{e51_4}{years passed in lower secondary}

\item{e51_5}{years passed in upper secondary}

\item{e51_6}{years passed in technical upper secondary}

\item{e51_7}{years passed in technical school}

\item{e51_7_1}{technical school requirements}

\item{e51_8}{years passed in magisterio/profesorado}

\item{e51_9}{years passed in university or similar}

\item{e51_10}{years passed in tertiary (non-university)}

\item{e51_11}{years passed in postgrade}

\item{e193}{attendance school}

\item{e49}{attendance school ever}
}
\description{
A function to calculate the highest level of education achieved
}
\details{
Disclaimer: El script no es un producto oficial de INE.
}
\examples{
\donttest{
toy_ech_2018 <- level_education(data = ech::toy_ech_2018)
}
}
