% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{organize_educ}
\alias{organize_educ}
\title{This function allows you to fix education variables from 2021}
\usage{
organize_educ(data, year, e49 = "e49", e579 = "e579", numero = "numero")
}
\arguments{
\item{data}{data.frame}

\item{year}{survey year}

\item{e49}{Variable name of e49}

\item{e579}{Variable name of e579}

\item{numero}{Variable name of numero}
}
\value{
data.frame
}
\description{
This function allows you to fix education variables from 2021
}
\seealso{
Other education: 
\code{\link{enrolled_school}()},
\code{\link{level_completion}()},
\code{\link{level_education}()},
\code{\link{years_of_schooling}()}
}
\concept{education}
