% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{organize_ht11}
\alias{organize_ht11}
\title{This function allows you to fix ht11 from 2013 to 2015}
\usage{
organize_ht11(data, year, ht11 = "ht11", numero = "numero")
}
\arguments{
\item{data}{data.frame}

\item{year}{survey year}

\item{ht11}{Variable name of ht11}

\item{numero}{Variable name of numero}
}
\value{
data.frame
}
\description{
This function allows you to fix ht11 from 2013 to 2015
}
\examples{
toy_ech_2018 <- organize_ht11(data = ech::toy_ech_2018, year = 2018)
}
\seealso{
Other income: 
\code{\link{basket_goods}()},
\code{\link{deflate}()},
\code{\link{income_constant_prices}()},
\code{\link{income_quantiles}()},
\code{\link{labor_income_per_capita}()},
\code{\link{labor_income_per_hour}()}
}
\concept{income}
