% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwelling.R
\name{overcrowding}
\alias{overcrowding}
\title{A function to calculate overcrowding in the household}
\usage{
overcrowding(data = ech::toy_ech_2018, ht19 = "ht19", d10 = "d10")
}
\arguments{
\item{data}{data.frame}

\item{ht19}{Number of individuals in the household}

\item{d10}{Number of rooms to sleep}
}
\value{
data.frame
}
\description{
A function to calculate overcrowding in the household
}
\examples{
toy_ech_2018 <- overcrowding(data = ech::toy_ech_2018)

}
