% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_microdata.R
\name{read_ech}
\alias{read_ech}
\title{Lee los archivos descargados de la ECH}
\usage{
read_ech(archivo)
}
\arguments{
\item{archivo}{cadena de texto con la ruta de la encuesta en formato zip/rar o sav/dat}
}
\value{
la encuesta ECH en formato tibble
}
\description{
Lee los archivos descargados de la ECH
}
\examples{
# datos de ejemplo en el paquete
read_ech(system.file(package = "ech", "extdata", "ech_2017_spss.rar"))

# descargando los datos
# download_ech(2017, "data-raw")
# read_ech("data-raw/ech_2017_spss.rar")
}
