% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organize.R
\name{to_ascii}
\alias{to_ascii}
\title{to_ascii}
\usage{
to_ascii(x, upper = TRUE)
}
\arguments{
\item{x}{a column}

\item{upper}{logic. Default TRUE}
}
\value{
vector
}
\description{
to_ascii
}
\examples{
\donttest{
d <- lapply(dic, to_ascii)
}
}
\seealso{
Other organize: 
\code{\link{organize_names}()}
}
\concept{organize}
