% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toy_ech_2018}
\alias{toy_ech_2018}
\title{Data of size 1000}
\format{A data frame with 1000 rows and 579 variables:
\describe{
  \item{numero}{household id}
  ...
}}
\source{
\url{http://www.ine.gub.uy/}
}
\usage{
toy_ech_2018
}
\description{
A dataset containing only 1000 raws of the ECH 2018
}
\keyword{datasets}
