% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toy_ech_2018_income}
\alias{toy_ech_2018_income}
\title{Data of 2018 incomes}
\format{
A data frame with 1000 rows and 9 variables:
\describe{
  \item{numero}{household id}
  \item{mes}{}
  \item{ht11}{}
  \item{ysvl}{}
  \item{ht13}{}
  \item{ht19}{}
  \item{dpto}{}
  \item{pesoano}{}
  \item{estred13}{}
  \item{anio}{}
  \item{region_4}{}
  ...
}
}
\source{
\url{http://www.ine.gub.uy/}
}
\usage{
toy_ech_2018_income
}
\description{
A dataset containing only 1000 raws of the ECH 2018 income variables
}
\details{
Disclaimer: This script is not an official INE product.
Aviso: El script no es un producto oficial de INE.
}
\keyword{datasets}
