% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/employment.R
\name{underemployment}
\alias{underemployment}
\title{This function allows you to identify underemployed people}
\usage{
underemployment(
  data = ech::toy_ech_2018,
  pobpcoac = "pobpcoac",
  f85 = "f85",
  f98 = "f98",
  f101 = "f101",
  f102 = "f102",
  f103 = "f103",
  f104 = "f104"
)
}
\arguments{
\item{data}{data.frame}

\item{pobpcoac}{Variable name of definition of population by activity status. Default: "pobpcoac"}

\item{f85}{Variable name of number of hours worked in the main job}

\item{f98}{Variable name of Number of hours worked at the secondary job}

\item{f101}{Variable name of reasons why you want another job}

\item{f102}{Variable name of want to work more hours}

\item{f103}{Variable name of are available to work more hours at this time}

\item{f104}{Variable name of reasons why you dont work more hours}
}
\value{
data.frame
}
\description{
This function allows you to identify underemployed people
}
\details{
Disclaimer: This script is not an official INE product.
Aviso: El script no es un producto oficial de INE.
}
\examples{
toy_ech_2018 <- underemployment(data = ech::toy_ech_2018)
}
\seealso{
Other employment: 
\code{\link{branch_ciiu}()},
\code{\link{employment_restrictions}()},
\code{\link{employment}()}
}
\concept{employment}
