% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unlabelled}
\alias{unlabelled}
\title{unlabelled}
\usage{
unlabelled(
  data = NULL,
  variable = NULL,
  by.x = NULL,
  by.y = NULL,
  ids = "numero",
  e30 = "e30",
  strata = "estred13",
  weights = "pesoano"
)
}
\arguments{
\item{data}{data frame}

\item{variable}{data frame column}

\item{by.x}{data frame column}

\item{by.y}{data frame column}

\item{ids}{ids}

\item{e30}{e30}

\item{strata}{strata}

\item{weights}{weights}
}
\value{
data.frame
}
\description{
unlabelled
}
\examples{
df <- unlabelled(data = ech::toy_ech_2018, variable = "pobre06")

}
