% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_microdata.R
\docType{data}
\name{unrarPath}
\alias{unrarPath}
\alias{.unrarPath}
\title{The known path for unrar or 7z}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
.unrarPath
}
\description{
The known path for unrar or 7z
}
\seealso{
Other utils: 
\code{\link{archive_extract}()},
\code{\link{dates_ech}()},
\code{\link{ech}},
\code{\link{unlabelled}()}
}
\concept{utils}
\keyword{datasets}
