% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{urls_ine}
\alias{urls_ine}
\title{Data urls_ine}
\format{
A data frame with 9 rows and 4 variables:
\describe{
  \item{yy}{date from 2011 to 2019}
  \item{md_sav}{url for microdata download}
  \item{upm_sav}{url for upm download}
  \item{dic}{url for dictionary download}
 }
}
\source{
\url{http://www.ine.gub.uy/}
}
\usage{
urls_ine
}
\description{
A dataset containing the urls of INE datasets and diccionaries
}
\details{
Disclaimer: This script is not an official INE product.
Aviso: El script no es un producto oficial de INE.
}
\keyword{datasets}
