% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/education.R
\name{years_of_schooling}
\alias{years_of_schooling}
\title{This function allows you to calculate the years of schooling}
\usage{
years_of_schooling(
  data = ech::toy_ech_2018,
  e193 = "e193",
  e51_2 = "e51_2",
  e51_3 = "e51_3",
  e51_4 = "e51_4",
  e51_5 = "e51_5",
  e51_6 = "e51_6",
  e51_7 = "e51_7",
  e51_7_1 = "e51_7_1",
  e51_8 = "e51_8",
  e51_9 = "e51_9",
  e51_10 = "e51_10",
  e51_11 = "e51_11",
  max_years = 22
)
}
\arguments{
\item{data}{data.frame}

\item{e193}{Variable name of attendance school}

\item{e51_2}{Variable name of years passed in primary}

\item{e51_3}{Variable name of years passed in special primary}

\item{e51_4}{Variable name of years passed in lower secondary}

\item{e51_5}{Variable name of years passed in upper secondary}

\item{e51_6}{Variable name of years passed in bachillerato tecnologico}

\item{e51_7}{Variable name of years passed in technical education}

\item{e51_7_1}{Variable name of technical education requirements}

\item{e51_8}{Variable name of years passed in magisterio/profesorado}

\item{e51_9}{Variable name of years passed in university or similar}

\item{e51_10}{Variable name of years passed in tertiary (non-university)}

\item{e51_11}{Variable name of years passed in postgrade}

\item{max_years}{Maximum years of schooling}
}
\value{
data.frame
}
\description{
This function allows you to calculate the years of schooling
}
\details{
Disclaimer: This script is not an official INE product.
Aviso: El script no es un producto oficial de INE.
}
\examples{
\donttest{
toy_ech_2018 <- years_of_schooling(data = ech::toy_ech_2018)
}
}
\seealso{
Other education: 
\code{\link{enrolled_school}()},
\code{\link{level_completion}()},
\code{\link{level_education}()},
\code{\link{organize_educ}()}
}
\concept{education}
