% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_geom}
\alias{add_geom}
\title{add_geom
See \code{geouy::\link[geouy]{add_geom}} for details.}
\value{
sf and data.frame object
}
\description{
add_geom
See \code{geouy::\link[geouy]{add_geom}} for details.
}
\keyword{internal}
