% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{age_groups}
\alias{age_groups}
\title{This function allows you to calculate age groups}
\usage{
age_groups(data = ech::toy_ech_2018, cut = c(0, 4, 11, 17, 24), e27 = "e27")
}
\arguments{
\item{data}{data.frame}

\item{cut}{breaks points to cut a numeric variable}

\item{e27}{Variable name of age}
}
\value{
data.frame
}
\description{
This function allows you to calculate age groups
}
\examples{
#' toy_ech_2018 <- age_groups(data = ech::toy_ech_2018, cut = c(0, 4, 11, 17, 24))
}
\seealso{
Other demographic: 
\code{\link{household_type}()}
}
\concept{demographic}
