% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/employment.R
\name{branch_ciiu}
\alias{branch_ciiu}
\title{Title}
\usage{
branch_ciiu(data = ech::toy_ech_2018, f72_2 = "f72_2", group = TRUE)
}
\arguments{
\item{data}{data.frame}

\item{f72_2}{ciiu code rev.4}

\item{group}{logical}
}
\value{
data.frame
}
\description{
Title
}
\examples{
\donttest{
toy_ech_2018 <- branch_ciiu(data = ech::toy_ech_2018)
}
}
