% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/employment.R
\name{branch_ciiu}
\alias{branch_ciiu}
\title{This function allows you to identify activity branches}
\usage{
branch_ciiu(
  data = ech::toy_ech_2018,
  f72_2 = "f72_2",
  group = TRUE,
  disaggregated = FALSE
)
}
\arguments{
\item{data}{data.frame}

\item{f72_2}{Variable name of ciiu code rev.4}

\item{group}{logical to define 12 or 18 categories, if FALSE code 18. Default: TRUE}

\item{disaggregated}{logical to define disaggregated branches or not. Default: FALSE}
}
\value{
data.frame
}
\description{
This function allows you to identify activity branches
}
\details{
Disclaimer: This script is not an official INE product.
Aviso: El script no es un producto oficial de INE.
}
\examples{
\donttest{
toy_ech_2018 <- branch_ciiu(data = ech::toy_ech_2018)
}
}
\seealso{
Other employment: 
\code{\link{employment_restrictions}()},
\code{\link{employment}()},
\code{\link{underemployment}()}
}
\concept{employment}
