% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cba_cbna_int}
\alias{cba_cbna_int}
\title{A dataset containing the CBA and CBNA for the Interior Urbano region}
\format{
A data frame with 234 rows and 4 variables:
\describe{
  \item{fecha}{date from 2001 to 2020}
  \item{cba_li}{CBA}
  \item{cbna}{CBNA}
  \item{cbt_lp}{CBT}
 }
}
\source{
\url{https://www.gub.uy/instituto-nacional-estadistica/}
}
\usage{
cba_cbna_int
}
\description{
A dataset containing the CBA and CBNA for the Interior Urbano region
}
\details{
Disclaimer: This script is not an official INE product.
Aviso: El script no es un producto oficial de INE.
}
\seealso{
Other dataset: 
\code{\link{cba_cbna_mdeo}},
\code{\link{cba_cbna_rur}},
\code{\link{dic}},
\code{\link{ipab_base2010_int}},
\code{\link{ipab_base2010_mdeo}},
\code{\link{ipab_base2010}},
\code{\link{ipc_base2010_int}},
\code{\link{ipc_base2010_mdeo}},
\code{\link{ipc_base2010}},
\code{\link{toy_ech_2018_income}},
\code{\link{toy_ech_2018}},
\code{\link{urls_ine}}
}
\concept{dataset}
\keyword{datasets}
