% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ciiu4}
\alias{ciiu4}
\title{A dataset containing Clasificación Industrial Internacional Uniforme Rev. 4 and 3}
\format{
A data frame with 738 rows and 3 variables:
\describe{
  \item{ciiu_4}{Code of Clasificación Industrial Internacional Uniforme Rev. 4}
  \item{description}{Description of Clasificación Industrial Internacional Uniforme Rev. 4}
  \item{ciiu_3}{Code of Clasificación Industrial Internacional Uniforme Rev. 3}
  ...
 }
}
\source{
\url{http://www.ine.gub.uy/}
}
\usage{
ciiu4
}
\description{
A dataset containing Clasificación Industrial Internacional Uniforme Rev. 4 and 3
}
\details{
Disclaimer: This script is not an official INE product.
Aviso: El script no es un producto oficial de INE.
}
\seealso{
Other dataset: 
\code{\link{cba_cbna_int}},
\code{\link{cba_cbna_mdeo}},
\code{\link{cba_cbna_rur}},
\code{\link{dic}},
\code{\link{ipab_base2010_int}},
\code{\link{ipab_base2010_mdeo}},
\code{\link{ipab_base2010}},
\code{\link{ipc_base2010_int}},
\code{\link{ipc_base2010_mdeo}},
\code{\link{ipc_base2010}},
\code{\link{toy_ech_2018_income}},
\code{\link{toy_ech_2018}},
\code{\link{urls_ine}}
}
\concept{dataset}
\keyword{datasets}
