% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dates_ech}
\alias{dates_ech}
\title{This function allows you to organize dates}
\usage{
dates_ech(data)
}
\arguments{
\item{data}{data frame with an 'yy' variable for the year, and a 'mm' variable for the month}
}
\value{
data.frame
}
\description{
This function allows you to organize dates
}
\seealso{
Other utils: 
\code{\link{archive_extract}()},
\code{\link{ech}},
\code{\link{unlabelled}()},
\code{\link{unrarPath}}
}
\concept{utils}
