% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{deflate}
\alias{deflate}
\title{deflate}
\usage{
deflate(base_month = NULL, base_year = NULL, ipc = "G", df_year = NULL)
}
\arguments{
\item{base_month}{baseline month}

\item{base_year}{baseline year}

\item{ipc}{General IPC ('G'), Montevideo IPC ('M') or Interior IPC ('I')}

\item{df_year}{ECH year}
}
\description{
deflate
}
\details{
Disclaimer: This script is not an official INE product.
Aviso: El script no es un producto oficial de INE.
}
\examples{
\donttest{
deflate(base_month = "06", base_year = "2016", df_year = "2018")
}
}
