% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{deflate}
\alias{deflate}
\title{deflate}
\usage{
deflate(base_month = NULL, base_year = NULL, ipc = "G", df_year = NULL)
}
\arguments{
\item{base_month}{mes base}

\item{base_year}{anio base}

\item{ipc}{IPC a nivel nacional ('G'), IPC para Montevideo ('M') e IPC para Interior ('I')}

\item{df_year}{anio del data frame}
}
\description{
deflate
}
\examples{
\donttest{
deflate(base_month = "06", base_year = "2016", df_year = "2016")
}
}
