% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{deflate}
\alias{deflate}
\title{This function allows you to calculate a deflator coefficient}
\usage{
deflate(
  base_month = NULL,
  base_year = NULL,
  index = "IPC",
  level = "G",
  df_year = NULL
)
}
\arguments{
\item{base_month}{baseline month}

\item{base_year}{baseline year}

\item{index}{IPC or IPAB}

\item{level}{General index ('G'), Montevideo index ('M') or Interior index ('I')}

\item{df_year}{ECH year}
}
\value{
vector
}
\description{
This function allows you to calculate a deflator coefficient
}
\details{
Disclaimer: This script is not an official INE product.
Aviso: El script no es un producto oficial de INE.
}
\seealso{
Other income: 
\code{\link{basket_goods}()},
\code{\link{income_constant_prices}()},
\code{\link{income_quantiles}()},
\code{\link{labor_income_per_capita}()},
\code{\link{labor_income_per_hour}()},
\code{\link{organize_ht11}()}
}
\concept{income}
