% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dic}
\alias{dic}
\title{A dataset containing variables names change of the ECH 2006-2018}
\format{
A data frame with 976 rows and 19 variables:
\describe{
  \item{codigos}{Code oh label}
  \item{descripcion}{Description of label}
  \item{modulo}{Module in the form 2017}
  \item{obs}{Observations}
  \item{unidad}{Level of variable household (H) individual (P) or general (G)}
  \item{var06}{ECH variables names 2006}
  \item{var07}{ECH variables names 2007}
  \item{var08}{ECH variables names 2008}
  \item{var09}{ECH variables names 2009}
  \item{var10}{ECH variables names 2010}
  \item{var11}{ECH variables names 2011}
  \item{var12}{ECH variables names 2012}
  \item{var13}{ECH variables names 2013}
  \item{var14}{ECH variables names 2014}
  \item{var15}{ECH variables names 2015}
  \item{var16}{ECH variables names 2016}
  \item{var17}{ECH variables names 2017}
  \item{var18}{ECH variables names 2018}
  \item{var19}{ECH variables names 2019}
  ...
}
}
\source{
\url{http://www.ine.gub.uy/}
}
\usage{
dic
}
\description{
A dataset containing variables names change of the ECH 2006-2018
}
\details{
Disclaimer: This script is not an official INE product.
Aviso: El script no es un producto oficial de INE.
}
\seealso{
Other dataset: 
\code{\link{cba_cbna_int}},
\code{\link{cba_cbna_mdeo}},
\code{\link{cba_cbna_rur}},
\code{\link{ciiu4}},
\code{\link{ipab_base2010_int}},
\code{\link{ipab_base2010_mdeo}},
\code{\link{ipab_base2010}},
\code{\link{ipc_base2010_int}},
\code{\link{ipc_base2010_mdeo}},
\code{\link{ipc_base2010}},
\code{\link{toy_ech_2018_income}},
\code{\link{toy_ech_2018}},
\code{\link{urls_ine}}
}
\concept{dataset}
\keyword{datasets}
