% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/employment.R
\name{employment}
\alias{employment}
\title{This function allows you to calculate the variables: PEA, PET, PO, PD}
\usage{
employment(data = ech::toy_ech_2018, pobpcoac = "pobpcoac")
}
\arguments{
\item{data}{data.frame with microdata}

\item{pobpcoac}{Variable name of definition of population by activity status. Default: "pobpcoac"}
}
\description{
This function allows you to calculate the variables: PEA, PET, PO, PD
}
\details{
Disclaimer: This script is not an official INE product.
Aviso: El script no es un producto oficial de INE.
}
\examples{
\donttest{
toy_ech_2018 <- employment(data = ech::toy_ech_2018, pobpcoac = "pobpcoac")
}
}
\seealso{
Other employment: 
\code{\link{branch_ciiu}()},
\code{\link{employment_restrictions}()},
\code{\link{underemployment}()}
}
\concept{employment}
\keyword{employment}
