% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/employment_rate.R
\name{employment_rate}
\alias{employment_rate}
\title{An employment rate function}
\usage{
employment_rate(data = df, pobcoac = "pobcoac")
}
\arguments{
\item{data}{data frame with microdata}

\item{pobcoac}{Definition of categories: "Ocupados", "Desocupados buscan trab. por 1a. vez", "Desocupados propiamente dichos", "Desocupados en seguro de paro"}
}
\description{
This function allows you to..
}
\examples{
employment_rate()
}
\keyword{employment}
