% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/employment.R
\name{employment_restrictions}
\alias{employment_restrictions}
\title{This function allows you to identify workers with employment restrictions}
\usage{
employment_restrictions(
  data = ech::toy_ech_2018,
  f82 = "f82",
  underemployment = "underemployment"
)
}
\arguments{
\item{data}{data.frame}

\item{f82}{Variable name of contribution to the pension fund}

\item{underemployment}{Variable name of underemployment}
}
\value{
data.frame
}
\description{
This function allows you to identify workers with employment restrictions
}
\details{
Disclaimer: This script is not an official INE product.
Aviso: El script no es un producto oficial de INE.
}
\examples{
toy_ech_2018 <- underemployment(data = ech::toy_ech_2018)
toy_ech_2018 <- employment_restrictions(data = toy_ech_2018)
}
\seealso{
Other employment: 
\code{\link{branch_ciiu}()},
\code{\link{employment}()},
\code{\link{underemployment}()}
}
\concept{employment}
