% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrolled_school.R
\name{enrolled_school}
\alias{enrolled_school}
\title{A function to estimate people enrolled in school}
\usage{
enrolled_school(data = ech::toy_ech_2018)
}
\arguments{
\item{data}{data.frame with necessary variables Defaults to ech.}
}
\description{
This function allows you to estimate people enrolled in school
}
\details{
Disclaimer: El script no es un producto oficial de INE.
}
\examples{
\donttest{
enrolled_school(data = ech::toy_ech_2018)
}
}
\keyword{education}
