% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrolled_school.R
\name{enrolled_school}
\alias{enrolled_school}
\title{A function to estimate people enrolled in school}
\usage{
enrolled_school(
  data = ech::toy_ech_2018,
  e193 = "e193",
  e197 = "e197",
  e201 = "e201",
  e212 = "e212",
  e215 = "e215",
  e218 = "e218",
  e221 = "e221",
  e224 = "e224"
)
}
\arguments{
\item{data}{data.frame with necessary variables Defaults to ech.}

\item{e193}{attendance school}

\item{e197}{attendance primary}

\item{e201}{attendance secondary}

\item{e212}{attendance technical education (non-university)}

\item{e215}{attendance magisterio}

\item{e218}{attendance university}

\item{e221}{attendance tertiary}

\item{e224}{attendance postgrade}
}
\description{
This function allows you to estimate people enrolled in school
}
\details{
Disclaimer: El script no es un producto oficial de INE.
}
\examples{
\donttest{
toy_ech_2018 <- enrolled_school(data = ech::toy_ech_2018)
}
}
\keyword{education}
