% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/education.R
\name{enrolled_school}
\alias{enrolled_school}
\title{This function allows you to calculate the people enrolled in school}
\usage{
enrolled_school(
  data = ech::toy_ech_2018,
  e27 = "e27",
  e193 = "e193",
  e197 = "e197",
  e201 = "e201",
  e212 = "e212",
  e215 = "e215",
  e218 = "e218",
  e221 = "e221",
  e224 = "e224"
)
}
\arguments{
\item{data}{data.frame with necessary variables Defaults to ech.}

\item{e27}{Variable name of age}

\item{e193}{Variable name of attendance school}

\item{e197}{Variable name of attendance primary}

\item{e201}{Variable name of attendance secondary}

\item{e212}{Variable name of attendance technical school (non-university)}

\item{e215}{Variable name of attendance magisterio}

\item{e218}{Variable name of attendance university}

\item{e221}{Variable name of attendance tertiary}

\item{e224}{Variable name of attendance postgrade}
}
\value{
data.frame
}
\description{
This function allows you to calculate the people enrolled in school
}
\details{
Disclaimer: This script is not an official INE product.
Aviso: El script no es un producto oficial de INE.
}
\examples{
\donttest{
toy_ech_2018 <- enrolled_school(data = ech::toy_ech_2018)
}
}
\seealso{
Other education: 
\code{\link{level_completion}()},
\code{\link{level_education}()},
\code{\link{years_of_schooling}()}
}
\concept{education}
\keyword{education}
