% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_cba_cbna}
\alias{get_cba_cbna}
\title{Title}
\usage{
get_cba_cbna(folder = tempdir(), sheet = NULL, region = NULL)
}
\arguments{
\item{folder}{temporal folder}

\item{sheet}{number of sheet}

\item{region}{Montevideo ("M"), Interior Urbano ("I"), Interior Rural ("R")}
}
\value{
data.frame
}
\description{
Title
}
