% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_cba_cbna}
\alias{get_cba_cbna}
\title{get_cba_cbna}
\usage{
get_cba_cbna(folder = tempdir(), sheet = NULL, region = NULL)
}
\arguments{
\item{folder}{temporal folder}

\item{sheet}{sheet number}

\item{region}{Montevideo ("M"), Interior Urbano ("I"), Interior Rural ("R")}
}
\value{
data.frame
}
\description{
This function allows you to get the CBA and CBNA data
}
\details{
Disclaimer: This script is not an official INE product.
Aviso: El script no es un producto oficial de INE.
}
\examples{
get_cba_cbna(folder = tempdir(), sheet = 1, region = "M")

}
