% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_cba_cbna}
\alias{get_cba_cbna}
\title{This function allows you to get the CBA and CBNA data}
\usage{
get_cba_cbna(folder = tempdir(), region, sheet = 1)
}
\arguments{
\item{folder}{temporal folder}

\item{region}{Montevideo ("M"), Interior Urbano ("I"), Interior Rural ("R")}

\item{sheet}{sheet number. Default 1.}
}
\value{
data.frame
}
\description{
This function allows you to get the CBA and CBNA data
}
\details{
Disclaimer: This script is not an official INE product.
Aviso: El script no es un producto oficial de INE.
}
\seealso{
Other dwnld_read: 
\code{\link{get_dictionary}()},
\code{\link{get_ipab_region}()},
\code{\link{get_ipab}()},
\code{\link{get_ipc_region}()},
\code{\link{get_ipc}()},
\code{\link{get_microdata}()},
\code{\link{read_microdata}()}
}
\concept{dwnld_read}
