% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_ciiu}
\alias{get_ciiu}
\title{get_ciiu}
\usage{
get_ciiu(folder = tempdir(), version = 4)
}
\arguments{
\item{folder}{temp folder}

\item{version}{by default the last ciiu version}
}
\description{
This function allows you to get the CIIU data
}
\details{
Disclaimer: This script is not an official INE product.
Aviso: El script no es un producto oficial de INE.
}
