% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_ciiu}
\alias{get_ciiu}
\title{This function allows you to get the CIIU data}
\usage{
get_ciiu(folder = tempdir(), version = 4)
}
\arguments{
\item{folder}{temp folder}

\item{version}{by default the last ciiu version}
}
\description{
This function allows you to get the CIIU data
}
\details{
Disclaimer: This script is not an official INE product.
Aviso: El script no es un producto oficial de INE.
}
\seealso{
Other dwnld_read: 
\code{\link{get_cba_cbna}()},
\code{\link{get_dictionary}()},
\code{\link{get_ipab}()},
\code{\link{get_ipc_region}()},
\code{\link{get_ipc}()},
\code{\link{get_microdata}()},
\code{\link{read_microdata}()}
}
\concept{dwnld_read}
