% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dictionary.R
\name{get_dictionary}
\alias{get_dictionary}
\title{This function allows you to download ECH dictionaries from INE website}
\usage{
get_dictionary(year = NULL, folder = tempdir())
}
\arguments{
\item{year}{allows download data from 2011 to 2019. Default the last year}

\item{folder}{Folder where are the files or be download}
}
\value{
unrar files from INE web and the respective data frame in tibble format
}
\description{
This function allows you to download ECH dictionaries from INE website
}
\details{
Disclaimer: This script is not an official INE product.
Aviso: El script no es un producto oficial de INE.
}
\seealso{
Other dwnld_read: 
\code{\link{get_cba_cbna}()},
\code{\link{get_ipab_region}()},
\code{\link{get_ipab}()},
\code{\link{get_ipc_region}()},
\code{\link{get_ipc}()},
\code{\link{get_microdata}()},
\code{\link{read_microdata}()}
}
\concept{dwnld_read}
