% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_estimation.R
\name{get_estimation_gini}
\alias{get_estimation_gini}
\title{get_estimation_gini}
\usage{
get_estimation_gini(
  data = ech::toy_ech_2018,
  variable = NULL,
  by = NULL,
  level = NULL,
  name = "estimacion"
)
}
\arguments{
\item{data}{ech data frame}

\item{variable}{income without rental value per capita deflated}

\item{by}{variable}

\item{level}{household or individual}

\item{name}{nombre}
}
\value{
table
}
\description{
get_estimation_gini
}
\examples{
\donttest{
df <- income_constant_prices(data = ech::toy_ech_2018, ipc = "R",
 base_month = "01", base_year = "2005")
get_estimation_gini(data = df, variable = "y_wrv_pc_d_r", level = "h")
}
}
