% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_estimation.R
\name{get_estimation_gini}
\alias{get_estimation_gini}
\title{This function allows you to estimate the Gini coeficient}
\usage{
get_estimation_gini(
  data = ech::toy_ech_2018,
  variable = NULL,
  by = NULL,
  level = NULL,
  ids = NULL,
  numero = "numero",
  estrato = NULL,
  pesoano = "pesoano",
  name = "estimacion"
)
}
\arguments{
\item{data}{ech data frame}

\item{variable}{income without rental value per capita deflated}

\item{by}{variable}

\item{level}{household or individual}

\item{ids}{upm}

\item{numero}{Variable name of household id}

\item{estrato}{Variable name of strata}

\item{pesoano}{Variable name of weights}

\item{name}{nombre}
}
\value{
table
}
\description{
This function allows you to estimate the Gini coeficient
}
\details{
Disclaimer: This script is not an official INE product.
Aviso: El script no es un producto oficial de INE.
}
\examples{
\donttest{
toy_ech_2018 <- income_constant_prices(data = ech::toy_ech_2018, ipc = "R",
 base_month = "01", base_year = "2005")
get_estimation_gini(data = toy_ech_2018, level = "h")
}
}
\seealso{
Other estimation: 
\code{\link{get_estimation_mean}()},
\code{\link{get_estimation_ratio}()},
\code{\link{get_estimation_total}()},
\code{\link{set_design}()}
}
\concept{estimation}
