% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_estimation.R
\name{get_estimation_gini}
\alias{get_estimation_gini}
\title{Title}
\usage{
get_estimation_gini(
  data = ech::toy_ech_2018,
  variable = ht11_svl_per_capita_deflate_gini,
  by.x = NULL,
  by.y = NULL,
  domain = NULL,
  level = NULL,
  name = "estimacion"
)
}
\arguments{
\item{data}{ech data frame}

\item{variable}{igreso sin valor locativo deflactado}

\item{by.x}{variable}

\item{by.y}{variable}

\item{domain}{subpoblacion}

\item{level}{nivel}

\item{name}{nombre}
}
\value{
table
}
\description{
Title
}
