% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_estimation.R
\name{get_estimation_mean}
\alias{get_estimation_mean}
\title{A function to estimate variables at universe level}
\usage{
get_estimation_mean(data = df, variable = NULL, by.x = NULL,
  by.y = NULL, domain = NULL, level = NULL, design = d)
}
\arguments{
\item{data}{data frame with ECH microdata}

\item{variable}{data frame column to estimate}

\item{by.x}{data frame column}

\item{by.y}{data frame column}

\item{domain}{subpopulation reference}

\item{level}{is household or individual}

\item{design}{survey design}
}
\value{
table
}
\description{
This function allows you to estimate variable.
}
\examples{
get_estimation(data = df)
}
\keyword{inference}
