% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_estimation.R
\name{get_estimation_mean}
\alias{get_estimation_mean}
\title{get_estimation_mean}
\usage{
get_estimation_mean(
  data = ech::toy_ech_2018,
  variable = NULL,
  by.x = NULL,
  by.y = NULL,
  domain = NULL,
  level = NULL,
  ids = "numero",
  e30 = "e30",
  strata = "estred13",
  weights = "pesoano",
  name = "estimacion"
)
}
\arguments{
\item{data}{data frame with ECH microdata}

\item{variable}{data frame column to estimate}

\item{by.x}{data frame column}

\item{by.y}{data frame column}

\item{domain}{subpopulation reference setted as logical}

\item{level}{is household ("h") or individual ("i").}

\item{ids}{ids}

\item{e30}{e30}

\item{strata}{strata}

\item{weights}{weights}

\item{name}{name for the estimation new column}
}
\value{
table
}
\description{
This function allows you to estimate mean variable at universe level.
}
\details{
Disclaimer: This script is not an official INE product.
Aviso: El script no es un producto oficial de INE.
}
\examples{
\donttest{
get_estimation_mean(variable = "pobre06", by.x = "dpto", level = "h")
}
}
\keyword{inference}
