% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_estimation.R
\name{get_estimation_ratio}
\alias{get_estimation_ratio}
\title{A function to estimate ratio variables at universe level}
\usage{
get_estimation_ratio(
  data = ech::toy_ech_2018,
  variable.x = NULL,
  variable.y = NULL,
  by.x = NULL,
  by.y = NULL,
  domain = NULL,
  level = NULL
)
}
\arguments{
\item{data}{data frame with ECH microdata}

\item{variable.x}{data frame column to estimate}

\item{variable.y}{data frame column to estimate}

\item{by.x}{data frame column}

\item{by.y}{data frame column}

\item{domain}{subpopulation reference setted as logical}

\item{level}{is household or individual}
}
\value{
table
}
\description{
This function allows you to estimate ratio variable.
}
\details{
Disclaimer: El script no es un producto oficial de INE.
}
\examples{
\donttest{
toy_ech_2018 <- employment(data = ech::toy_ech_2018, pobpcoac = "pobpcoac")
get_estimation_ratio(data = toy_ech_2018, variable.x = "po", variable.y = "pea", level = "i")
}
}
\keyword{inference}
