% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_estimation.R
\name{get_estimation_ratio}
\alias{get_estimation_ratio}
\title{get_estimation_ratio}
\usage{
get_estimation_ratio(
  data = ech::toy_ech_2018,
  variable.x = NULL,
  variable.y = NULL,
  by.x = NULL,
  by.y = NULL,
  domain = NULL,
  level = NULL,
  ids = NULL,
  numero = "numero",
  estred13 = "estred13",
  pesoano = "pesoano",
  name = "estimacion"
)
}
\arguments{
\item{data}{data frame with ECH microdata}

\item{variable.x}{data frame column to estimate}

\item{variable.y}{data frame column to estimate}

\item{by.x}{data frame column}

\item{by.y}{data frame column}

\item{domain}{subpopulation reference setted as logical}

\item{level}{is household ("h") or individual ("i")}

\item{ids}{ids}

\item{numero}{household id}

\item{estred13}{strata}

\item{pesoano}{weights}

\item{name}{name for the estimation new column}
}
\value{
table
}
\description{
This function allows you to estimate ratio variable at universe level.
}
\details{
Disclaimer: This script is not an official INE product.
Aviso: El script no es un producto oficial de INE.
}
\examples{
\donttest{
toy_ech_2018 <- employment(data = ech::toy_ech_2018, pobpcoac = "pobpcoac")
get_estimation_ratio(data = toy_ech_2018, variable.x = "po", variable.y = "pea", level = "i")
}
}
\keyword{inference}
