% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_estimation.R
\name{get_estimation_total}
\alias{get_estimation_total}
\title{A function to estimate total variables at universe level}
\usage{
get_estimation_total(
  data = ech::toy_ech_2018,
  variable = NULL,
  by.x = NULL,
  by.y = NULL,
  domain = NULL,
  level = NULL
)
}
\arguments{
\item{data}{data frame with ECH microdata}

\item{variable}{data frame column to estimate}

\item{by.x}{data frame column}

\item{by.y}{data frame column}

\item{domain}{subpopulation reference}

\item{level}{is household ("h") or individual ("i").}
}
\value{
table
}
\description{
This function allows you to estimate total variable.
}
\details{
Disclaimer: El script no es un producto oficial de INE.
}
\examples{
\donttest{
get_estimation_total(variable = "pobre06", by.x = "dpto", level = "h")
}
}
\keyword{inference}
