% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_ipab}
\alias{get_ipab}
\title{This function allows you to get the IPAB (Indice de precios de alimentos y bebidas) data}
\usage{
get_ipab(folder = tempdir(), sheet = 1)
}
\arguments{
\item{folder}{temporal folder}

\item{sheet}{sheet number. Default 1}
}
\value{
data.frame
}
\description{
This function allows you to get the IPAB (Indice de precios de alimentos y bebidas) data
}
\details{
Disclaimer: This script is not an official INE product.
Aviso: El script no es un producto oficial de INE.
}
\seealso{
Other dwnld_read: 
\code{\link{get_cba_cbna}()},
\code{\link{get_dictionary}()},
\code{\link{get_ipab_region}()},
\code{\link{get_ipc_region}()},
\code{\link{get_ipc}()},
\code{\link{get_microdata}()},
\code{\link{read_microdata}()}
}
\concept{dwnld_read}
