% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_ipab_region}
\alias{get_ipab_region}
\title{This function allows you to get the IPAB (Indice de precios de alimentos y bebidas) data}
\usage{
get_ipab_region(folder = tempdir(), region, sheet = 1)
}
\arguments{
\item{folder}{temporal folder}

\item{region}{Montevideo ("M"), Interior Urbano ("I")}

\item{sheet}{sheet number. Default 1}
}
\value{
data.frame
}
\description{
This function allows you to get the IPAB (Indice de precios de alimentos y bebidas) data
}
\seealso{
Other dwnld_read: 
\code{\link{get_cba_cbna}()},
\code{\link{get_dictionary}()},
\code{\link{get_ipab}()},
\code{\link{get_ipc_region}()},
\code{\link{get_ipc}()},
\code{\link{get_microdata}()},
\code{\link{read_microdata}()}
}
\concept{dwnld_read}
