% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_ipc}
\alias{get_ipc}
\title{This function allows you to get the IPC data}
\usage{
get_ipc(folder = tempdir())
}
\arguments{
\item{folder}{ruta temporal para descargar el archivo}
}
\description{
This function allows you to get the IPC data
}
\details{
Disclaimer: This script is not an official INE product.
Aviso: El script no es un producto oficial de INE.
}
\seealso{
Other dwnld_read: 
\code{\link{get_cba_cbna}()},
\code{\link{get_dictionary}()},
\code{\link{get_ipab_region}()},
\code{\link{get_ipab}()},
\code{\link{get_ipc_region}()},
\code{\link{get_microdata}()},
\code{\link{read_microdata}()}
}
\concept{dwnld_read}
