% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_ipc_region}
\alias{get_ipc_region}
\title{get_ipc_region}
\usage{
get_ipc_region(folder = tempdir(), region = "M", sheet = NULL)
}
\arguments{
\item{folder}{temporal folder}

\item{region}{Montevideo ("M") or Interior ("I")}

\item{sheet}{sheet number}
}
\value{
data.frame
}
\description{
This function allows you to get the IPC data
}
\details{
Disclaimer: This script is not an official INE product.
Aviso: El script no es un producto oficial de INE.
}
\examples{
\donttest{
get_ipc_region(folder = tempdir(), region = "M")
}

}
