% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_microdata.R
\name{get_marco}
\alias{get_marco}
\alias{get_sampling_frame}
\title{This unction allows you to download and read the sampling frame from INE website}
\usage{
get_marco(year = NULL, folder = tempdir(), toR = TRUE)

get_sampling_frame(year = NULL, folder = tempdir(), toR = TRUE)
}
\arguments{
\item{year}{allows download data of 2011, 2004, 1996 or 1985. Default 2011}

\item{folder}{Folder where are the files or be download}

\item{toR}{write data frame in R format and delete download file and unpack files}
}
\value{
unrar files from INE web and the respective data frame in tibble format
}
\description{
This unction allows you to download and read the sampling frame from INE website
}
\details{
Disclaimer: This script is not an official INE product.
Aviso: El script no es un producto oficial de INE.
}
\seealso{
Other dwnld_read: 
\code{\link{get_cba_cbna}()},
\code{\link{get_ciiu}()},
\code{\link{get_dictionary}()},
\code{\link{get_ipab}()},
\code{\link{get_ipc_region}()},
\code{\link{get_ipc}()},
\code{\link{get_microdata}()},
\code{\link{read_microdata}()}
}
\concept{dwnld_read}
