% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_microdata.R
\name{get_microdata}
\alias{get_microdata}
\title{Download and read ECH from INE website}
\usage{
get_microdata(year = NULL, folder = NULL, toR = TRUE)
}
\arguments{
\item{year}{allows download data from 2011 to 2018. Default the last year}

\item{folder}{Folder where are the files or be download}

\item{toR}{write data frame in R format and delete download file and unpack files}
}
\value{
unrar files from INE web and the respective data frame in tibble format
}
\description{
Download and read ECH from INE website
}
\examples{
# download and read all surveys available
# Download and read ECH 2017
get_microdata(year = "2014", folder = "/home/calcita/Escritorio/pruebaech/", toR = FALSE)
}
