% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_microdata.R
\name{get_microdata}
\alias{get_microdata}
\title{This function allows you to download and read ECH from INE website}
\usage{
get_microdata(year = NULL, folder = tempdir(), toR = TRUE)
}
\arguments{
\item{year}{allows download data from 2011 to 2019. Default the last year}

\item{folder}{Folder where are the files or be download}

\item{toR}{write data frame in R format and delete download file and unpack files}
}
\value{
unrar files from INE web and the respective data frame in tibble format
}
\description{
This function allows you to download and read ECH from INE website
}
\details{
Disclaimer: This script is not an official INE product.
Aviso: El script no es un producto oficial de INE.
}
\seealso{
Other dwnld_read: 
\code{\link{get_cba_cbna}()},
\code{\link{get_dictionary}()},
\code{\link{get_ipab_region}()},
\code{\link{get_ipab}()},
\code{\link{get_ipc_region}()},
\code{\link{get_ipc}()},
\code{\link{read_microdata}()}
}
\concept{dwnld_read}
