% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gini.R
\name{gini}
\alias{gini}
\title{gini}
\usage{
gini(data = ech::toy_ech_2018, df_year = 2018)
}
\arguments{
\item{data}{data frame with ECH microdata}

\item{df_year}{year of the ech data.frame}
}
\description{
gini
}
