% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gini.R
\name{gini}
\alias{gini}
\title{gini}
\usage{
gini(data = df, weights = pesoano, domain = NULL)
}
\arguments{
\item{data}{data frame with ECH microdata}

\item{weights}{ponderation variable}

\item{domain}{dominio de interes}
}
\description{
gini
}
\examples{
gini(df, pesoano)
}
