% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwelling.R
\name{homeownership}
\alias{homeownership}
\title{Title}
\usage{
homeownership(data = ech::toy_ech_2018, d8_1 = "d8_1")
}
\arguments{
\item{data}{data.frame}

\item{d8_1}{tenencia de la vivienda}
}
\value{
data.frame
}
\description{
Title
}
\examples{
toy_ech_2018 <- homeownership(data = ech::toy_ech_2018)

}
