% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/household_type.R
\name{household_type}
\alias{household_type}
\title{household_type
This function allows you to calculate the household type for each household in the survey. A household is composed of one or more people who occupy a housing unit.}
\usage{
household_type(
  data = ech::toy_ech_2018,
  e26 = "e26",
  e27 = "e27",
  e30 = "e30",
  colname = "tipo_hogar"
)
}
\arguments{
\item{data}{data frame with ECH microdata}

\item{e26}{sex}

\item{e27}{age}

\item{e30}{householder}

\item{colname}{custom name for the new variable}
}
\value{
data.frame
}
\description{
household_type
This function allows you to calculate the household type for each household in the survey. A household is composed of one or more people who occupy a housing unit.
}
\details{
Disclaimer: This script is not an official INE product.
Aviso: El script no es un producto oficial de INE.
}
\examples{
\donttest{
toy_ech_2018 <- household_type(data = ech::toy_ech_2018)
}
}
\keyword{household_type}
