% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/household_type.R
\name{household_type}
\alias{household_type}
\title{This function allows you to calculate the household type for each household in the survey. A household is composed of one or more people who occupy a housing unit.}
\usage{
household_type(
  data = ech::toy_ech_2018,
  numero = "numero",
  e26 = "e26",
  e27 = "e27",
  e30 = "e30"
)
}
\arguments{
\item{data}{data frame with ECH microdata}

\item{numero}{Variable name of household id}

\item{e26}{Variable name of sex}

\item{e27}{Variable name of age}

\item{e30}{Variable name of householder}
}
\value{
data.frame
}
\description{
This function allows you to calculate the household type for each household in the survey. A household is composed of one or more people who occupy a housing unit.
}
\details{
Disclaimer: This script is not an official INE product.
Aviso: El script no es un producto oficial de INE.
}
\examples{
\donttest{
toy_ech_2018 <- household_type(data = ech::toy_ech_2018)
}
}
\seealso{
Other demographic: 
\code{\link{age_groups}()}
}
\concept{demographic}
\keyword{household_type}
