% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwelling.R
\name{housing_conditions}
\alias{housing_conditions}
\title{housing_conditions}
\usage{
housing_conditions(data = ech::toy_ech_2018, c2 = "c2", c3 = "c3", c4 = "c4")
}
\arguments{
\item{data}{data.frame}

\item{c2}{material predominante en paredes externas}

\item{c3}{material predominante en techos}

\item{c4}{material predominante en pisos}
}
\value{
data.frame
}
\description{
housing_conditions
}
\examples{
toy_ech_2018 <- housing_conditions(data = ech::toy_ech_2018)

}
