% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwelling.R
\name{housing_deprivation}
\alias{housing_deprivation}
\title{A function to calculate housing status}
\usage{
housing_deprivation(
  data = ech::toy_ech_2018,
  ht19 = "ht19",
  d9 = "d9",
  d10 = "d10",
  d11 = "d11",
  d12 = "d12",
  d13 = "d13",
  d16 = "d16",
  d18 = "d18",
  d19 = "d19",
  c2 = "c2",
  c3 = "c3",
  c4 = "c4",
  quintil = "quintil",
  region_4 = "region_4"
)
}
\arguments{
\item{data}{data frame con las variables necesarias de ech}

\item{ht19}{cantidad de personas sin servicio domestico}

\item{d9}{cantidad de habitaciones para dormir}

\item{d10}{cantidad de habitaciones para dormir}

\item{d11}{origen del agua}

\item{d12}{llegada del agua a la vivienda}

\item{d13}{servicio sanitario}

\item{d16}{evacuacion del servicio sanitario}

\item{d18}{fuente de energia para iluminar}

\item{d19}{lugar para cocinar}

\item{c2}{material predominante en paredes externas}

\item{c3}{material predominante en techo}

\item{c4}{material predominante en pisos}

\item{quintil}{numero de quintil de ingresos}

\item{region_4}{region}
}
\value{
data.frame
}
\description{
This function allows you to calculate the housing status
}
\details{
Disclaimer: El script no es un producto oficial de INE.
}
\examples{
\donttest{
toy_ech_2018 <- income_constant_prices(data = ech::toy_ech_2018)
toy_ech_2018 <- income_quantiles(data = toy_ech_2018)
toy_ech_2018 <- housing_deprivation(data = toy_ech_2018)
}
}
