% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwelling.R
\name{housing_situation}
\alias{housing_situation}
\title{Title}
\usage{
housing_situation(
  data = ech::toy_ech_2018,
  c5_1 = "c5_1",
  c5_2 = "c5_2",
  c5_3 = "c5_3",
  c5_4 = "c5_4",
  c5_5 = "c5_5",
  c5_6 = "c5_6",
  c5_7 = "c5_7",
  c5_8 = "c5_8",
  c5_9 = "c5_9",
  c5_10 = "c5_10",
  c5_11 = "c5_11",
  c5_12 = "c5_12"
)
}
\arguments{
\item{data}{data.frame}

\item{c5_1}{roof dampness}

\item{c5_2}{roof leaks}

\item{c5_3}{cracked walls}

\item{c5_4}{broken doors or windows}

\item{c5_5}{cracked floors}

\item{c5_6}{caida de revoque en paredes}

\item{c5_7}{cielos rasos desprendidos}

\item{c5_8}{poca luz solar}

\item{c5_9}{escasa ventilacion}

\item{c5_10}{se inunda cuando llueve}

\item{c5_11}{peligro de derrumbe}

\item{c5_12}{humedades en los cimientos}
}
\value{
data.frame
}
\description{
Title
}
