% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/income.R
\name{income_constant_prices}
\alias{income_constant_prices}
\title{income_constant_prices}
\usage{
income_constant_prices(
  data = ech::toy_ech_2017_income,
  base_month = 6,
  base_year = 2017,
  df_year = base_year,
  mes = "mes",
  ht11 = "ht11",
  ysvl = "YSVL",
  ht13 = "ht13",
  ht19 = "ht19"
)
}
\arguments{
\item{data}{data frame with ECH microdata}

\item{base_month}{mes base}

\item{base_year}{anio base}

\item{df_year}{anio ech (default anio base)}

\item{mes}{mes}

\item{ht11}{ht11}

\item{ysvl}{ysvl}

\item{ht13}{ht13}

\item{ht19}{ht19}
}
\description{
Household income constant prices
}
\details{
Disclaimer: El script no es un producto oficial de INE.
}
\examples{
\donttest{
toy_ech_2017_income <- income_constant_prices(data = ech::toy_ech_2017_income)
}
}
