% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/income_constant_prices.R
\name{income_constant_prices}
\alias{income_constant_prices}
\title{income_constant_prices}
\usage{
income_constant_prices(data = df, base.month = 6, base.year = 2017,
  mes = mes, ht11 = ht11, ysvl = ysvl, ht13 = ht13, ht19 = ht19)
}
\description{
Household income constant prices
}
