% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/income_constant_prices.R
\name{income_constant_prices}
\alias{income_constant_prices}
\title{income_constant_prices}
\usage{
income_constant_prices(
  data = ech::toy_ech_2017_income,
  base.month = 6,
  base.year = 2017,
  mes = "mes",
  ht11 = "ht11",
  ysvl = "YSVL",
  ht13 = "ht13",
  ht19 = "ht19"
)
}
\arguments{
\item{data}{data frame with ECH microdata}

\item{base.month}{mes base}

\item{base.year}{anio base}

\item{mes}{mes}

\item{ht11}{ht11}

\item{ysvl}{ysvl}

\item{ht13}{ht13}

\item{ht19}{ht19}
}
\description{
Household income constant prices
}
