% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/income_quantiles.R
\name{income_quantiles}
\alias{income_quantiles}
\title{income_quantiles}
\usage{
income_quantiles(
  data = ech::toy_ech_2017_income,
  quantile = 5,
  weights = "pesoano",
  income = "ht11_per_capita_deflate"
)
}
\arguments{
\item{data}{data.frame}

\item{quantile}{cuantiles: quintiles (5) o deciles (10)}

\item{weights}{ponderation variable}

\item{income}{Name of the income variables. Default: "ht11_per_capita_deflate"}
}
\description{
Household Income Quantiles
}
\examples{
\donttest{
a <- income_constant_prices()
b <- income_quantiles(data = a)
}
}
