% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/income_quantiles.R
\name{income_quantiles}
\alias{income_quantiles}
\title{income_quantiles}
\usage{
income_quantiles(
  data = ech::toy_ech_2017_income,
  quantile = 5,
  weights = "pesoano"
)
}
\arguments{
\item{data}{data.frame}

\item{quantile}{cuantiles: quintiles (5) o deciles (10)}

\item{weights}{ponderation variable}
}
\description{
Household Income Quantiles
}
\examples{
\donttest{
income_quantiles()
}
}
