% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/income.R
\name{income_quantiles}
\alias{income_quantiles}
\title{This function allows you to calculate the Household Income Quantiles}
\usage{
income_quantiles(
  data = ech::toy_ech_2018,
  quantile = 5,
  weights = "pesoano",
  income = "y_pc_d"
)
}
\arguments{
\item{data}{data.frame}

\item{quantile}{Variable name of quintil (5) or decil (10). Default: 5}

\item{weights}{Variable name of ponderation variable. Default: "pesoano"}

\item{income}{Variable name of income constant price. Default: "y_pc_d"}
}
\value{
data.frame
}
\description{
This function allows you to calculate the Household Income Quantiles
}
\details{
Disclaimer: This script is not an official INE product.
Aviso: El script no es un producto oficial de INE.
}
\examples{
\donttest{
toy_ech_2018 <- income_constant_prices(data = ech::toy_ech_2018)
toy_ech_2018 <- income_quantiles(data = toy_ech_2018)
}
}
\seealso{
Other income: 
\code{\link{basket_goods}()},
\code{\link{deflate}()},
\code{\link{income_constant_prices}()},
\code{\link{labor_income_per_capita}()},
\code{\link{labor_income_per_hour}()},
\code{\link{organize_ht11}()}
}
\concept{income}
