% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/income_quantiles.R
\name{income_quintiles}
\alias{income_quintiles}
\title{income_quintiles}
\usage{
income_quintiles(data = df,
  ht11_per_capita_deflate = ht11_per_capita_deflate, design = d,
  quantile = 5, weights = pesoano)
}
\description{
Household Income Quantiles
}
