% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ipc_base2010}
\alias{ipc_base2010}
\title{A dataset containing the IPC base 2010}
\format{
A data frame with 990 rows and 5 variables:
\describe{
  \item{fecha}{date from 1937 to 2019}
  \item{indice}{IPC}
  \item{mensual}{value of IPC}
  \item{acum_ano}{acumulated IPC}
  \item{acum_12_meses}{acumulated IPC last 12 month}
 }
}
\source{
\url{http://www.ine.gub.uy/}
}
\usage{
ipc_base2010
}
\description{
A dataset containing the IPC base 2010
}
\details{
Disclaimer: This script is not an official INE product.
Aviso: El script no es un producto oficial de INE.
}
\seealso{
Other dataset: 
\code{\link{cba_cbna_int}},
\code{\link{cba_cbna_mdeo}},
\code{\link{cba_cbna_rur}},
\code{\link{ciiu4}},
\code{\link{dic}},
\code{\link{ipab_base2010_int}},
\code{\link{ipab_base2010_mdeo}},
\code{\link{ipab_base2010}},
\code{\link{ipc_base2010_int}},
\code{\link{ipc_base2010_mdeo}},
\code{\link{toy_ech_2018_income}},
\code{\link{toy_ech_2018}},
\code{\link{urls_ine}}
}
\concept{dataset}
\keyword{datasets}
