% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ipc_base2010_int}
\alias{ipc_base2010_int}
\title{A dataset containing the IPC base 2010 only for the Interior region}
\format{
A data frame with 120 rows and 2 variables:
\describe{
  \item{fecha}{date from 2011 to 2019}
  \item{indice}{IPC}
 }
}
\source{
\url{https://www.gub.uy/instituto-nacional-estadistica/}
}
\usage{
ipc_base2010_int
}
\description{
A dataset containing the IPC base 2010 only for the Interior region
}
\details{
Disclaimer: This script is not an official INE product.
Aviso: El script no es un producto oficial de INE.
}
\seealso{
Other dataset: 
\code{\link{cba_cbna_int}},
\code{\link{cba_cbna_mdeo}},
\code{\link{cba_cbna_rur}},
\code{\link{dic}},
\code{\link{ipab_base2010_int}},
\code{\link{ipab_base2010_mdeo}},
\code{\link{ipab_base2010}},
\code{\link{ipc_base2010_mdeo}},
\code{\link{ipc_base2010}},
\code{\link{toy_ech_2018_income}},
\code{\link{toy_ech_2018}},
\code{\link{urls_ine}}
}
\concept{dataset}
\keyword{datasets}
