% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/income.R
\name{labor_income_per_capita}
\alias{labor_income_per_capita}
\title{This function allows you to calculate the labor income per capita}
\usage{
labor_income_per_capita(
  data = ech::toy_ech_2018,
  numero = "numero",
  pobpcoac = "pobpcoac",
  g126_1 = "g126_1",
  g126_2 = "g126_2",
  g126_3 = "g126_3",
  g126_4 = "g126_4",
  g126_5 = "g126_5",
  g126_6 = "g126_6",
  g126_7 = "g126_7",
  g126_8 = "g126_8",
  g127_3 = "g127_3",
  g128_1 = "g128_1",
  g129_2 = "g129_2",
  g130_1 = "g130_1",
  g131_1 = "g131_1",
  g133_1 = "g133_1",
  g133_2 = "g133_2",
  g134_1 = "g134_1",
  g134_2 = "g134_2",
  g134_3 = "g134_3",
  g134_4 = "g134_4",
  g134_5 = "g134_5",
  g134_6 = "g134_6",
  g134_7 = "g134_7",
  g134_8 = "g134_8",
  g135_3 = "g135_3",
  g136_1 = "g136_1",
  g137_2 = "g137_2",
  g138_1 = "g138_1",
  g139_1 = "g139_1",
  g141_1 = "g141_1",
  g141_2 = "g141_2",
  g142 = "g142",
  g144_1 = "g144_1",
  g144_2_1 = "g144_2_1",
  g144_2_3 = "g144_2_3",
  g144_2_4 = "g144_2_4",
  g144_2_5 = "g144_2_5"
)
}
\arguments{
\item{data}{data frame}

\item{numero}{Variable name of household id}

\item{pobpcoac}{Variable name of definition of population by activity status}

\item{g126_1}{Variable name of net salary}

\item{g126_2}{Variable name of commissions, incentives, overtime payment, fringe benefits}

\item{g126_3}{Variable name of non-surrendering expenses}

\item{g126_4}{Variable name of tips}

\item{g126_5}{Variable name of annual complementary salary}

\item{g126_6}{Variable name of vacation pay}

\item{g126_7}{Variable name of delayed payments}

\item{g126_8}{Variable name of transportation tickets}

\item{g127_3}{Variable name of received food or drink}

\item{g128_1}{Variable name of received food tickets}

\item{g129_2}{Variable name of received housing or accommodation}

\item{g130_1}{Variable name of another type of compensation}

\item{g131_1}{Variable name of received another type of supplement paid by the employer}

\item{g133_1}{Variable name of the right to cultivate goods for own-consumption}

\item{g133_2}{Variable name of the right to cultivate goods for own-consumption (amount received from the sale)}

\item{g134_1}{Variable name of net salary}

\item{g134_2}{Variable name of commissions, incentives, overtime payment, fringe benefits}

\item{g134_3}{Variable name of non-surrendering expenses}

\item{g134_4}{Variable name of tips}

\item{g134_5}{Variable name of annual complementary salary}

\item{g134_6}{Variable name of vacation pay}

\item{g134_7}{Variable name of delayed payments}

\item{g134_8}{Variable name of transportation tickets}

\item{g135_3}{Variable name of received food or drink}

\item{g136_1}{Variable name of received food tickets}

\item{g137_2}{Variable name of received housing or accommodation}

\item{g138_1}{Variable name of another type of compensation}

\item{g139_1}{Variable name of received another type of supplement paid by the employer}

\item{g141_1}{Variable name of the right to cultivate goods for own-consumption}

\item{g141_2}{Variable name of the right to cultivate goods for own-consumption (amount received from the sale)}

\item{g142}{Variable name of withdrawals for business household expenses you have or had}

\item{g144_1}{Variable name of collected products for own consumption (non-agricultural worker)}

\item{g144_2_1}{Variable name of collected products for own consumption (non-agricultural worker)}

\item{g144_2_3}{Variable name of collected products for own consumption (non-agricultural worker)}

\item{g144_2_4}{Variable name of collected products for own consumption (non-agricultural worker)}

\item{g144_2_5}{Variable name of collected products for own consumption (non-agricultural worker)}
}
\value{
data.frame
}
\description{
This function allows you to calculate the labor income per capita
}
\details{
Disclaimer: This script is not an official INE product.
Aviso: El script no es un producto oficial de INE.
}
\examples{
\donttest{
toy_ech_2018 <- labor_income_per_capita(data = ech::toy_ech_2018)
}
}
\seealso{
Other income: 
\code{\link{basket_goods}()},
\code{\link{deflate}()},
\code{\link{income_constant_prices}()},
\code{\link{income_quantiles}()},
\code{\link{labor_income_per_hour}()},
\code{\link{organize_ht11}()}
}
\concept{income}
