% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/income.R
\name{labor_income_per_hour}
\alias{labor_income_per_hour}
\title{labor_income_per_hour}
\usage{
labor_income_per_hour(
  data = ech::toy_ech_2018,
  numero = "numero",
  f85 = "f85",
  pobpcoac = "pobpcoac",
  pt4 = "pt4",
  base_month = NULL,
  base_year = NULL,
  mes = "mes"
)
}
\arguments{
\item{data}{data frame}

\item{numero}{identificador de hogar}

\item{f85}{horas trabajadas por semana}

\item{pobpcoac}{condicion de actividad economica}

\item{pt4}{total de ingresos por trabajo}

\item{base_month}{mes base}

\item{base_year}{anio base}

\item{mes}{mes}
}
\value{
data.frame
}
\description{
labor_income_per_hour
}
\details{
Disclaimer: El script no es un producto oficial de INE.
}
\examples{
\donttest{
base <- ech::toy_ech_2018
toy_ech_2018 <- labor_income_per_hour(data = base, base_month = "06", base_year = "2018")
}

}
