% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/education.R
\name{level_completion}
\alias{level_completion}
\title{level_completion}
\usage{
level_completion(
  data = ech::toy_ech_2018,
  e197 = "e197",
  e197_1 = "e197_1",
  e201 = "e201",
  e51_4 = "e51_4",
  e51_5 = "e51_5",
  e51_6 = "e51_6"
)
}
\arguments{
\item{data}{data.frame}

\item{e197}{Variable name of attends primary school}

\item{e197_1}{Variable name of completed primary}

\item{e201}{Variable name of attends secondary}

\item{e51_4}{Variable name of years passed in lower secondary}

\item{e51_5}{Variable name of years passed in upper secondary}

\item{e51_6}{Variable name of years passed in technical upper secondary}
}
\value{
data.frame
}
\description{
This function allows you to calculate the level of school completion
}
\details{
Disclaimer: This script is not an official INE product.
Aviso: El script no es un producto oficial de INE.
}
\examples{
\donttest{
toy_ech_2018 <- level_completion(data = ech::toy_ech_2018)
}

}
