% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organize_labels.R
\name{organize_labels}
\alias{organize_labels}
\title{Funcion para etiquetar las bases de la ECH con referencia al 2017.}
\usage{
organize_labels(df, year, level = "individual")
}
\arguments{
\item{df}{type de microdatos de la ECH}

\item{year}{numeric reference year of the df. Avaible 2006 to 2018}

\item{level}{(string) aclaracion sobre si la base a etiquetar es de tipo "household", "h", "individual", "i", "hyp"}
}
\description{
Funcion para etiquetar las bases de la ECH con referencia al 2017.
}
\details{
Disclaimer: El script no es un producto oficial de INE.
}
\examples{
df <- organize_labels(df = toy_ech_2018, year = 2018, level="h")
}
