% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwelling.R
\name{overcrowding}
\alias{overcrowding}
\title{Title}
\usage{
overcrowding(data = ech::toy_ech_2018, ht19 = "ht19", d10 = "d10")
}
\arguments{
\item{data}{data.frame}

\item{ht19}{cantidad integrantes}

\item{d10}{bedrooms}
}
\value{
data.frame
}
\description{
Title
}
\examples{
toy_ech_2018 <- overcrowding(data = ech::toy_ech_2018)

}
