% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwelling.R
\name{overcrowding}
\alias{overcrowding}
\title{This function allows you to calculate overcrowding in the household}
\usage{
overcrowding(data = ech::toy_ech_2018, ht19 = "ht19", d10 = "d10")
}
\arguments{
\item{data}{data.frame}

\item{ht19}{Variable name of umber of individuals in the household}

\item{d10}{Variable name of number of rooms to sleep}
}
\value{
data.frame
}
\description{
This function allows you to calculate overcrowding in the household
}
\details{
Disclaimer: This script is not an official INE product.
Aviso: El script no es un producto oficial de INE.
}
\examples{
toy_ech_2018 <- overcrowding(data = ech::toy_ech_2018)
}
\seealso{
Other dwelling: 
\code{\link{housing_conditions}()},
\code{\link{housing_deprivation}()},
\code{\link{housing_situation}()},
\code{\link{housing_tenure}()}
}
\concept{dwelling}
