% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe operator
See \code{dplyr::\link[dplyr]{\%>\%}} for details.}
\usage{
lhs \%>\% rhs
}
\description{
Pipe operator
See \code{dplyr::\link[dplyr]{\%>\%}} for details.
}
\keyword{internal}
