% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poverty.R
\name{poverty}
\alias{poverty}
\title{This function allows you to calculate poor and indigent people or household}
\usage{
poverty(
  data = ech::toy_ech_2018,
  scale = 0.8,
  region_4 = "region_4",
  dpto = "dpto",
  ht11 = "ht11",
  ht19 = "ht19",
  numero = "numero"
)
}
\arguments{
\item{data}{data.frame}

\item{scale}{equivalency scale}

\item{region_4}{Variable name of region. Default: region_4}

\item{dpto}{Variable name of departamento. Default: dpto}

\item{ht11}{Variable name of income. Default: ht11}

\item{ht19}{Variable name of number of individuals in the household. Default: ht19}

\item{numero}{household id}
}
\value{
data.frame
}
\description{
This function allows you to calculate poor and indigent people or household
}
\details{
Disclaimer: This script is not an official INE product.
Aviso: El script no es un producto oficial de INE.
}
\examples{
toy_ech_2018 <- poverty(data = ech::toy_ech_2018)
}
\seealso{
Other poverty: 
\code{\link{integrated_poverty_measure}()},
\code{\link{unsatisfied_basic_needs}()}
}
\concept{poverty}
