% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_microdata.R
\name{read_microdata}
\alias{read_microdata}
\title{Read ECH from a local folder}
\usage{
read_microdata(path = NULL)
}
\arguments{
\item{path}{Folder where are the files or be download}
}
\value{
an object called df
}
\description{
Read ECH from a local folder
}
\examples{
read_microdata(folder = "/home/usuarie/Escritorio/HyP_2017_Terceros.sav")
}
