% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_microdata.R
\name{read_microdata}
\alias{read_microdata}
\title{Read ECH from a local folder}
\usage{
read_microdata(folder = NULL, filename = NULL, format = "rds")
}
\arguments{
\item{folder}{Folder where are the files or be download}

\item{filename}{File name}

\item{format}{allows read data in rds, rdata, sav and dta}
}
\value{
unrar files from INE web and the respective data frame in tibble format
}
\description{
Read ECH from a local folder
}
\examples{
read_microdata()
}
