% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_microdata.R
\name{read_microdata}
\alias{read_microdata}
\title{Read ECH from a local folder}
\usage{
read_microdata(path = NULL)
}
\arguments{
\item{path}{Folder where are the files or be download}
}
\value{
an object called df
}
\description{
Read ECH from a local folder
}
\details{
Disclaimer: El script no es un producto oficial de INE.
}
