% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_design.R
\name{set_design}
\alias{set_design}
\title{A function to set survey design}
\usage{
set_design(data = data, ids = 1, strata = estred13,
  weights = pesoano)
}
\arguments{
\item{data}{data frame with ECH microdata}

\item{ids}{survey phase}

\item{strata}{survey strata}

\item{weights}{ponderation variable}

\item{level}{is household or individual}
}
\value{
d
}
\description{
This function allows you to set the survey desing
}
\examples{
set_design(data = d)
}
\keyword{design}
