% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_design.R
\name{set_design}
\alias{set_design}
\title{set_design
This function allows you to set the survey desing}
\usage{
set_design(
  data = ech::toy_ech_2018,
  level = "i",
  ids = "numero",
  strata = "estred13",
  weights = "pesoano"
)
}
\arguments{
\item{data}{data frame with ECH microdata}

\item{level}{is household ("h") or individual ("i")}

\item{ids}{variables specifying cluster ids from largest level to smallest level (leaving the argument empty, NULL, 1, or 0 indicate no clusters).}

\item{strata}{variable specifying strata.}

\item{weights}{variable specifying weights (inverse of probability).}
}
\value{
a list
}
\description{
set_design
This function allows you to set the survey desing
}
\details{
Disclaimer: This script is not an official INE product.
Aviso: El script no es un producto oficial de INE.
}
\examples{
\donttest{
set_design(data = ech::toy_ech_2018, level = "h")
}
}
\keyword{design}
