% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_design.R
\name{set_design}
\alias{set_design}
\title{A function to set survey design}
\usage{
set_design(data = ech::toy_ech_2018, level = "i")
}
\arguments{
\item{data}{data frame with ECH microdata}

\item{level}{is household ("h") or individual ("i")}
}
\value{
d
}
\description{
This function allows you to set the survey desing
}
\examples{
set_design(data = ech::toy_ech_2018)
}
\keyword{design}
