% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organize.R
\name{to_ascii}
\alias{to_ascii}
\title{to_ascii}
\usage{
to_ascii(x, upper = T)
}
\arguments{
\item{x}{a column}

\item{upper}{logic}
}
\value{
vector
}
\description{
to_ascii
}
\examples{
\donttest{
d <- lapply(dic, to_ascii)
}
}
