% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toy_ech_2018_income}
\alias{toy_ech_2018_income}
\title{A dataset containing only 1000 raws of the ECH 2018 income variables}
\format{
A data frame with 1000 rows and 9 variables:
\describe{
  \item{numero}{household id}
  \item{mes}{}
  \item{ht11}{}
  \item{ysvl}{}
  \item{ht13}{}
  \item{ht19}{}
  \item{dpto}{}
  \item{pesoano}{}
  \item{estred13}{}
  \item{anio}{}
  \item{region_4}{}
  ...
}
}
\source{
\url{https://www.gub.uy/instituto-nacional-estadistica/}
}
\usage{
toy_ech_2018_income
}
\description{
A dataset containing only 1000 raws of the ECH 2018 income variables
}
\details{
Disclaimer: This script is not an official INE product.
Aviso: El script no es un producto oficial de INE.
}
\seealso{
Other dataset: 
\code{\link{cba_cbna_int}},
\code{\link{cba_cbna_mdeo}},
\code{\link{cba_cbna_rur}},
\code{\link{dic}},
\code{\link{ipab_base2010_int}},
\code{\link{ipab_base2010_mdeo}},
\code{\link{ipab_base2010}},
\code{\link{ipc_base2010_int}},
\code{\link{ipc_base2010_mdeo}},
\code{\link{ipc_base2010}},
\code{\link{toy_ech_2018}},
\code{\link{urls_ine}}
}
\concept{dataset}
\keyword{datasets}
