% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unlabelled}
\alias{unlabelled}
\title{unlabelled}
\usage{
unlabelled(data = NULL)
}
\arguments{
\item{data}{data frame}
}
\value{
data.frame
}
\description{
unlabelled
}
\examples{
df <- unlabelled(data = ech::toy_ech_2018)

}
