% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unlabelled}
\alias{unlabelled}
\title{This function allows you to labelled variables}
\usage{
unlabelled(data = NULL)
}
\arguments{
\item{data}{data frame}
}
\value{
data.frame
}
\description{
This function allows you to labelled variables
}
\examples{
df <- unlabelled(data = ech::toy_ech_2018)
}
\seealso{
Other utils: 
\code{\link{archive_extract}()},
\code{\link{dates_ech}()}
}
\concept{utils}
