% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{urls_ine}
\alias{urls_ine}
\title{A dataset containing the urls of INE datasets and diccionaries}
\format{
A data frame with 9 rows and 4 variables:
\describe{
  \item{yy}{date from 2011 to 2019}
  \item{md_sav}{url for microdata download}
  \item{upm_sav}{url for upm download}
  \item{dic}{url for dictionary download}
 }
}
\source{
\url{https://www.gub.uy/instituto-nacional-estadistica/}
}
\usage{
urls_ine
}
\description{
A dataset containing the urls of INE datasets and diccionaries
}
\details{
Disclaimer: This script is not an official INE product.
Aviso: El script no es un producto oficial de INE.
}
\seealso{
Other dataset: 
\code{\link{cba_cbna_int}},
\code{\link{cba_cbna_mdeo}},
\code{\link{cba_cbna_rur}},
\code{\link{dic}},
\code{\link{ipab_base2010_int}},
\code{\link{ipab_base2010_mdeo}},
\code{\link{ipab_base2010}},
\code{\link{ipc_base2010_int}},
\code{\link{ipc_base2010_mdeo}},
\code{\link{ipc_base2010}},
\code{\link{toy_ech_2018_income}},
\code{\link{toy_ech_2018}}
}
\concept{dataset}
\keyword{datasets}
